[mubi] Weaken mubi*_t type
This patch was originally motivated by the irritation of casting
between bit vectors and the enum type. But it turns out there's a
better reason for doing it: the IEEE spec is very vague about whether
a variable of an enum type can contain values other than the enum
values and, if so, what happens. (For anyone who wants to read along,
the Enumerations section is 6.19 of IEEE 1800-2017, and the only
discussion of this values is at the end of 6.19.5, where an
"out-of-range" value is assigned: whatever that means!)

In practice, I suspect that any sensible tool will just treat e.g. the
old mubi4_t as a "logic [3:0]", together with a pair of helpful named
values. However, I'm not sure it's actually guaranteed by the
standard, nor was the exact behaviour of a mubi4_t'(...) when the
inputs are bogus.

Rather than worrying about such things, let's just use bit
vectors (since that's how we're thinking about these possibly-slippery
objects anyway!). What we'd probably *really* like is akin to
Haskell's "newtype"s, where you'd have an n-bit type that's
interconvertible with an n-bit vector, but needs explicit casts.
Unfortunately, SystemVerilog doesn't support that sort of thing.

Existing design code shouldn't need any changes, although it will now
have some unnecessary type-casts. And we can now be sure that an
enthusiastic technically-standards-compliant synthesis tool won't undo
all our hard work!

The only existing DV code that needs tweaking is in
entropy_src_env_cfg, where we genuinely want the enum type because we
want to use its name() method. This looks a bit odd to me: if we don't
want to allow invalid encodings, it's probably easier to use a bit,
which we'd translate to MuBi in the driver. But that's out of scope
for this PR.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>