[rtl] Handle integrity errors properly for bus devices with no regs
The way we normally handle malformed TL accesses in "foo_reg_top" is
that we detect the error, steer the access to the register adapter,
and pass it a flag saying that the access should be responded to with
an error. This is really neat, but (unsurprisingly!) doesn't work if
there are no registers.

This commit tweaks things so that we instantiate a tlul_err_resp if
there are no registers on the interface. This always responds with an
error, so we can just steer bad accesses to it to get everything
working.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>