[dv] Only run registers through one csr_rw sequence at once
This is an issue up for blocks with multiple device interfaces. The
default behaviour of run_csr_vseq is to grab all of the registers
defined for the block. The run_tl_intg_err_vseq task runs in parallel
across the different ral_names, which means we do run_csr_vseq("rw")
for each of the interfaces in parallel.

It turns out that the csr_rw sequence isn't really "re-entrant". If
you've got two copies running at once then you can see a register R
with thread 0 writing it and thread 1 reading it at the same time.
This causes a UVM warning (and causes the test to fail).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>