[build] give one arg for each Variadic macro
To build with the -pedantic flag set without warnings, we need to pass
one argument to all variadic macros. This commit addresses all the ones
in the project that bazel builds so far. Later this will be important if
we want to endable building under other compilers than gcc.

Ran clang_format on the c files I edit in this commit.

Signed-off-by: Drew Macrae <drewmacrae@google.com>