[flash_ctrl] Remove oob handling altogether
The current oob handling is not complete.  It only handles an upper bound
check but does not check the lower bound.  This means the original issue
that caused this error can still happen as long as someone supplies a
lower address (such as 0x0).

After talking it through with the @a-will, the utility of this feature
just seems very minor.  It also complicates things by pulling a system
address map into a peripheral, which really does not seem like the right
way to go about this.

As a result, I have removed this feature altogether, and will instead rely
on software to make sure the addresses being fed to flash_ctrl are within
bounds.

Fixes #9444

Signed-off-by: Timothy Chen <timothytim@google.com>