[ entropy_src, dv ] Support for all phases of entropy generation
Now supports conditioning bypass, boot phase, startup phase & continuous phase.

- This commit adds functionality to the entropy_source environment to predicte the
current phase of entropy_src operation: BOOT, STARTUP or CONTINUOUS. Tracking
this phase is essential both for scoreboarding and for planning sequences.

- RNG Sequence generation is redesigned to assign certain AST-RNG outputs to
particular entropy_src output seeds.  Each seed is tracked by a particular "seed
index", which can be used to predict what phase the entropy source will be in
when it receives that seed.  Knowing the phase is important, as it also indicates
how much RNG data must be provided to generate the this seed..

- The progression from one phase to another can be halted by health test failures.
Therefore this commit introduces a stub of a function to predict whether a particular
RNG sub-sequence will fail a health-check.  This function is currently unimplemented.
For the purposes of sequence generation, it is not important to know which health test
will fail, only that _some_ health test will fail.  Thus it is envisioned that
the stub that is currently in entropy_src_base_vseq.sv will eventually be constructed
from smaller functions, one for each type of health check.
  - Note: these fine-grain health check functions will be needed for both
    scoreboarding and sequence generation purposes.

- Includes various configuration registers which will be useful for controlling
the size of the health-check windows, via CSRs.

Signed-off-by: Martin Lueker-Boden <martin.lueker-boden@wdc.com>
Co-authored-by: weicaiyang <49293026+weicaiyang@users.noreply.github.com>