[mask_rom] Perform basic AST initialization
Copy the configuration from OTP unconditionally and then poll the
ast_init_done flag until it is set.

Future commits will make this initialization more flexible. In
particular we need to skip AST configuration if OTP is not yet
programmed and also we need to add a timeout to the polling.

Note that the open source AST model does not validate the
configuration and ast_init_done is hardcoded in the RTL to 1.

Signed-off-by: Michael Munday <mike.munday@lowrisc.org>