[ottf,tests] Migrate crt_test to use OTTF ISRs.
The crt_test checks the operations in the testing `rom_ext`'s `flash_crt.S`.
Because of this, it does not run within the context of the OTTF.
However, the `flash_crt.S` does set the `mtvec` pointer, so ISRs must be
present in case an unexpected interrupt is fired.

As described in #8015, we plan to deprecate the default ISR handlers in
`sw/device/lib/handler.<c,h>` favor of the OTTF ISRs so that only on a
single set of default testing ISRs must be maintained. Therefore, this
commit migrates the crt_test to make use of the OTTF ISRs. Additionally,
this commit makes it possible to use the OTTF ISRs without requiring the
OTTF be linked into the test image too.

Signed-off-by: Timothy Trippel <ttrippel@google.com>