[sw] Pass on native C/C++ compiler to build opentitantool
Mundane is a dependency for opentitantool and fails to build with gcc-11.
Specify C/C++ compiler to use e.g. with the following command:

```
CC_FOR_BUILD=gcc-10 CXX_FOR_BUILD=g++-10 ./meson_init.sh -f
```

(This is similar to issue #6978. And identical to the building with
the signer tool.)

Change-Id: I387f6a155ca4022077ab878b81bd2334fd0d5fd0