[rom_ctrl] Tweak checker counter to stop at top of ROM
This means that once we've finished our ROM check, the address from
the checker will be 1fff (rather than zero). This should have no
effect except that if someone somehow manages to convince the ROM mux
to take the address from the checker then the subverted TL access will
respond with the top of ROM. The previous commit ensures that this
will have invalid integrity bits so the fetch will kill boot.

Before this change, if an attack like that succeeded then it would
allow replacing a fetched instruction with whatever we have at the
reset vector, which might be more exploitable.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>