[rom_ctrl] Add an alert that spots glitches on the checker counter
We don't care about glitches to the counter when we're still checking
things (because exploiting one of them would need a pre-image attack
on SHA3!), but we might care about the counter changing when we're in
the Done state.

Theoretically, it doesn't matter what the counter address is in this
state, but if someone manages to subvert the mux they might be able to
use it to generate a bogus address to respond to a TL access. However,
if they do that then we'll see a negedge of counter_done on the
following cycle and can raise an alert.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>