[otbn,rtl] Fix dmem_addr_bus calculation with nonzero base address
The dmem_addr_bus signal grabs the index of the 32-bit data word
that's being written. We can't infer this from the output of the sram
adapter because that expects 256-bit data words and encodes the
"offset" within the word in the write mask.

The bug was that we were including some of the upper bits of the TL
address in the checksum. These are zero if OTBN's registers start at
zero, but might be other values in general. Conveniently, we ensure
that the memory windows are naturally aligned, so we can calculate the
offset within the window (the number we want!) by just masking out all
bits greater than vbits(DmemSizeByte).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>