[otbn,dv] Read and write LOAD_CHECKSUM when writing a binary
This augments the load_elf_over_bus() task (used 50% of the time to
load binaries into a test) so that it adds reads/writes for the
LOAD_CHECKSUM register in the middle of loading the binary.

The scoreboard then predicts the value seen, which lets us test the
CRC checksum functionality.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>