[otbn,dv,doc] Split out functional coverage documentation
This is almost 1000 lines, compared to 136 lines in the rest of the
document. Split it out to avoid making the DV document itself
unreadable.

The page rename (index.md -> _index.md) is because we need to make the
dv directory into a "branch bundle" in Hugo terminology. A "leaf
bundle" wouldn't allow other pages in the same directory.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>