[otbn,dv] Improvements to external CSR coverage
This strengthens quite a few of the coverpoints that we'd
defined (when squinting at them, I noticed that they weren't really
testing anything!).

It also rejigs the documentation slightly to give a section per CSR.
The previous format with them in a bulleted list made it hard to say
everything we need to say.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>