[sw/boot_rom] Rename "boot_rom" to "test_rom".
The ROM code in `sw/device/boot_rom/` is used purely for testing purposes,
and is not designed to be used as the production mask ROM. The mask ROM
code lives in `sw/device/silicon_creator/mask_rom`. To make this more clear,
the `sw/device/boot_rom` ia renamed and relocated to
`sw/device/lib/testing/test_rom`.

This fixes #9154.

Signed-off-by: Timothy Trippel <ttrippel@google.com>