[fpga] Fix splice CI export files step.
The upload process defined in ci/upload-artifacts-template.yml uses an
rsync command that keeps track of the upstream dependencies of the CI
step. This commit updates the splice script so that it generates the new
bitstream using a new filename so that it can be picked up in the upload
script.

Other changes: add additional error checking to the splice script.

Signed-off-by: Miguel Osorio <miguelosorio@google.com>