[ibex] Enable bit-manipulation extension
This vendors in the latest version of Ibex supporting the OTEarlGrey
bitmanip configuration and enables this config in Earl Grey. This adds
support for the Zba, Zbb, Zbc and Zbs bit-manipulation sub-extensions
ratified in v.1.0.0 of the RISC-V Bit-Manipulation ISA Extension, as
well as support for the non-ratified sub-extensions Zbf, Zbp, Zbr,
Zbt part of the draft v.0.93 spec.

This resolves lowRISC/OpenTitan#2708.

Update lowrisc_ibex to lowRISC/ibex@410ffd34

Update code from upstream repository
https://github.com/lowRISC/ibex.git to revision
410ffd349da1d4c13f77b208bb3b93adb154fac0

* [bitmanip, doc] Update info on bitmanip support and area numbers
  (Pirmin Vogel)
* [syn] Add missing package dependency (Pirmin Vogel)
* [rtl] Remove redundant comments in decoder (Pirmin Vogel)
* [bitmanip] Add new configuration option for OpenTitan Earl Grey
  (Pirmin Vogel)

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>