[aes/rtl] Overlap state register clearing with unmasking final output
Previously, the state register clearing has always been performed in the
same cycle in which also the final output register was written. This is
perfectly suitable for unmasked implementations or when using fully
combinational S-Boxes e.g. masked Canright.

However, DOM S-Boxes take multiple clock cycles to evaluate and provide
valid output in the fifth clock cycle. This output then propgates
through the rest of the cipher core and is ultimately forwarded to the
output registers/mode logic (both unmasked). To cover potential
SCA leakage related to this unmasking, the state register should be
cleared with PRD simultaneously to the unmasking (happening before the
clock edge). The write enable needs thus to be asserted one clock
cycle earlier, i.e., in the fourth cycle.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>