[aes/rtl] Only update masking PRNG in first clock cycle of each round
With the improved DOM S-Boxes it is sufficient to update the masking
PRNG in just one cycle per round. Updating it in any clock cycle other
than the last cycle of the round is better in terms of SCA as the PRD
consumed by the DOM S-Boxes in the first clock cycle can then be
guaranteed to be stable (no transient effects from the non-linear layer
at the PRNG output propagate into the S-Boxes).

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>