[sw/silicon_creator] Avoid double-writes when invalidating boot_data entries
Currently, we invalidate previous `boot_data` entries by writing 0 to
their `identifier` fields. This scheme, however, is not compatible with
ECC and scrambling because writing a different value to a previously
written flash word results in flash_ctrl errors in subsequent reads.

This commit adds a new field (`is_valid`) to `boot_data_t` to be able to
invalidate previous `boot_data` entries without modifying previously
written flash words. This field is skipped when a `boot_data` entry is
written and is only written when invalidating previous entries.

Signed-off-by: Alphan Ulusoy <alphan@google.com>