[entropy_src, dv] New RNG failure seqeunces
In order to improve the testing of health checks this commit introduces
a new entropy_src_base_rng_seq.  This sequence is an extension of the
push_pull_indefinite_host_seq, but it provides new methods to control
(and predict) the randomization of the RNG data.

There are now three modes of operation for the RNG sequence:
- Typical: Of the three modes this is the highest entropy. In this
  base class, this typical entropy is ideal however a derived class
  could be made to generate weaker entropy
- Soft failure: the entropy of the RNG sequence is degraded either
  through bias or correlations in the output stream.  In the base
  class however, the soft failure mode is the same as typical.
- Hard failure: In this mode one or more bits becomes "stuck",
  which is a complete failure of the RNG stream.

Entry into the soft or hard failure modes is random.
There are two vairables hard_mtbf, and soft_mtbf (mean time before
failure) which control average time before failing into one of
these states.  If the parent vseq identifies that the RNG
sequence has encountered a failure (perhaps through an alert)
it can call the reset_rng sequence.

Signed-off-by: Martin Lueker-Boden <martin.lueker-boden@wdc.com>