[otbn,dv] Avoid spurious errors from multi-error testcases
As an example, consider the bnsid-2 testcase. This was causing
failures in the otbn_multi_err test. The testcase boils down to this
instruction:

   bn.sid x1, 0(x2)

where the call stack is empty (so the read of x1 underflows) and x2
equals 1 (so the store address is invalid).

The problem is that if the call stack really hasn't ever had any
values (true if bnsid-2 comes first) then it will respond with X's.
That's fine, except that this is used as an index into the big-number
register file, which means that the rdata and associated error bits
come out as X. What's more, we didn't actually have any intialised
data in the big-number register file anyway, so we'd have got Xs
whatever the index was.

This is a problem because (via the fatal_err signal) it causes the
controller state to go to X and eventually causes us to send an X in
the imem request signal, which causes an assertion failure.

Avoid the problem by storing zero to the call stack before
popping it again (which avoids passing X as an index to the big-number
register file) and then clearing w0.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>