[rom_ctrl,dv] Correct alert field prediction in cip_base_scoreboard
This fixes a regression failure in rom_ctrl_tl_intg_err. The rom_ctrl
testbench can see the problem because rom_ctrl has two TL interfaces,
meaning that concurrent accesses can spot when we mess up the timing.

As well as fixing the timing, change things so that we don't call
update_tl_alert_field_prediction in response to integrity errors on
the D side (because that won't cause an alert in the block).

Signed-off-by: Prajwala Puttappa <prajwalaputtappa@lowrisc.org>
Co-authored-by: Rupert Swarbrick <rswarbrick@lowrisc.org>