[prim_xilinx] Replace KEEP with DONT_TOUCH attributes
Previously, we have been using KEEP attributes to prevent synthesis
optimizations across these primitives to ensure that FI and SCA
countermeasures are not optimized away. However, according to the
Vivado Design Suite User Guide: Synthesis (UG901), the KEEP attribute
is commonly used in conjunction with timing constraints but
doesn't force place and route to keep the signal. Instead, the
DONT_TOUCH attribute is recommended for this purpose (see Chapter 2):

"Unlike KEEP and KEEP_HIERARCHY, DONT_TOUCH is forward-annotated to
place and route to prevent logic optimization."

In addition, the KEEP/DONT_TOUCH attributes are not supported on ports
of modules and instead should be applied to the module itself.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>