[aes] Fix wrapper and add basic Verilator testbench
The purpose of this wrapper is to enable FI experiments on synthesized
netlists. Over time, it got out of sync with the AES RTL. This commit
re-aligns the wrapper with the current AES RTL and adds a basic
Verilator testbench to quickly verify the operation of the wrapper and
simplify future re-alignment.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>