[otbn,dv] Only run the exp_end_addr check for some runs
The expected end address check wasn't quite working now that we had a
key sideload interface. The problem is that if you read from one of
the sideload key WSRs when there's no key, the operation immediately
stops with an error. That's fine, but we won't stop at the address we
expected.

To allow this check to run at least some of the time (giving us more
confidence in the rig), we define a new absent_key_allowed flag. If
it's true, we use the normal key sideload sequence and disable the
check. If it's false, we grab control of the sideload sequencer and
run a specialised version where the key might change, but it always
has a value. In this case, we can assert the end address again.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>