[prim_assert,dv] Use if condition in assert_init
In this context, an if condition and assert statement exhibit the same
behavior. For some strange reason, the fix made in #9017 for Xcelium
(addition of #0 delay) worked for IP level testbenches but not the chip
level. The chip level tests returned the same error (race condition
causing the signal value to reflect an X, failing the assertion). This
may very well be a tool bug, where Xcelium is sampling the preponed
value of the signal when it shouldn't, but this seems to not be
reproducible in a standalone testcase (given that IP level testbenches
work fine). So in the interest of time, we just switch this out to a if
condition, which has an added advantage of begin immune to assert
control system tasks.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>