[otbn] Squash big-number read errors if the index is unknown
To trigger this event, run something like the bnmovr-1 test at the
start of a simulation. It does the following:

    addi    x2, x0, 100
    bn.movr x2, x1

Here, the second instruction underflows the empty call stack, meaning
that the bignum source register for the move is undefined. (The
destination is also bogus in this case, but that doesn't really
matter)

Without this change, the bignum register file's rd_data_err_o signal
depends on what happens to be in the (uninitialised) call stack and
the (also uninitialised) bignum register file. In RTL simulation, this
will be X. In silicon, it could be anything :-)

Architecturally, we didn't actually load anything so we want to squash
the error response.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>