[otbn,dv] Avoid "double recov alert" problem
Some sequences of operations could cause a problem where we ran two
back-to-back operations, each of which causing a recoverable alert,
where the second operation finished while before the handshake of the
first alert finished. If this happened, we would essentially drop the
second alert, causing the scoreboard to complain.

The fix is to expose the "waiting for an alert" state from the
scoreboard and wait in the vseq for that, then for any following
recoverable alert handshake to finish. This way, we can't start the
second operation too early.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>