[otbn,dv] Clear the URND running flag when starting an operation
When writing the "WSR cleanup" commit 04984f2, I didn't bother to
reset URND when starting an operation on the basis that it would sort
itself out when the EDN data came in. Unfortunately, I forgot that
we're also tracking whether we're stalled waiting for that data.

The result is that sequences with more than one operation back-to-back
fail because the ISS steams ahead (causing an INSN_CNT mismatch).

Fix that.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>