[dif/alert_handler] Refactor class config struct.
A goal of #9899 is to provide an additional DIF to provide a fine grain
class configuration mechnism, rather than having to configure the entire
alert handler in a single shot. This commit refactors the class config
struct, and DIF code that uses it, to enable future development of a DIF
that can configure a single alert handler class (i.e., the escalation
protocol for said class).

Note, this commit temporarily disables some alert_handler unit tests
until the class configuration code is refactored into its own DIF. This
will be contained in a following commit.

This partially addresses a task in #9899.

Signed-off-by: Timothy Trippel <ttrippel@google.com>