[aes] Add lockable bit to force PRNG reseeding upon touching the key
This commit adds a configuration option to force the reseeding of the
PRNGs upon providing a new key. To allow this new configuration bit to
be locked, a new auxiliary control register has to be added as the
regtool doesn't allow locking individual fields of a register but just
entire registers.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>