[dif/alert_handler] Add ping timer config DIF.
This commit adds a DIF to configure, lock, and enable the ping timer
feature of the alert handler. This DIF allows more fine grain
configuration of the alert handler than the previous implementation
that configured all alert handler setting in a single shot.

Additionally, this commit fixes a bug in the checking of the REGWEN lock
for the ping timer. Previously, the wrong CSR (the enable instead of the
REGWEN) was checked.

This addresses a task in #9899.

Signed-off-by: Timothy Trippel <ttrippel@google.com>