[dif/alert_handler] Add X macros to simplify DIF code.
The alert handler has a list of classes and local alerts that are
converted to enums, that are used in several case statement. This commit
adds X macros to improve the maintainability (in the case that the
hardware changes and additional local alerts or classes are added), and
readability of the alert handler DIF code.

This commit was added in response to a review comment in #10040.

Signed-off-by: Timothy Trippel <ttrippel@google.com>