[ entropy_src, dv ] Clean shutdown sequences
The entropy_src has a unique feature: Under certain conditions it must _stop_
honoring requests for more data.   This can happen for instance if it is
configured for very strict health check thresholds.  Furthermore,
it is possible for the testbench to instruct this DUT to adhere to completely
unrealistic thresholds, and in such cases it is also impossible for the testbench
to tell the entropy_src to relax.  So in these cases, no CSRNG data will likely be
released for the remainder of the test.

(Such conditions can also appear in directed alert tests.)

These conditions are directly at odds with the push-pull PULL sequences which expect
every request to be acknowledged once REQ is asserted.  In such cases, the
push_pull_host_seq's will not gracefully exit.  Furthermore, even after the sequence
has been aborted, the push_pull_monitor will continue to object because until both
REQ and ACK have been cleared.   This leads to interminable simulations.

This commit solves this problem in two steps:
1. The push_pull_indefinite_host_seq now has an option for a "hard" stop
which will interrupt and halt the sequence, even if a sequence item is incomplete.
2. The testbench adds a second _auxilliary_ reset for the CSRNG pull agent.
This reset is only applied at the very end of the sequence, and _only_ to the CSRNG
VIF.  (Applying a global testbench reset at the end of the sim, both alters the state of the
DUT and for some simulators it also causes errors in the RNG push agent.)
  - This later step is only needed for the unit test environment.  In the top-level environment
    the entropy_src would be connected directly to the CSRNG (no monitor).

Signed-off-by: Martin Lueker-Boden <martin.lueker-boden@wdc.com>