[pwrmgr,dv] Connect up port properly in bind statement
The sw_rst_req_i signal became a mubi4_t in commit d7f00ea49f but we
forgot to update the bindings. Amusingly, this actually switches the
polarity of the input (because MuBi4True = 4'hA and MuBi4False = 4'h5,
taking the bottom bit gives the opposite of the boolean that we
intended!).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>