[otbn,dv] Fix how we override the sideload sequence
The previous code was using 'disable fork' to kill our special
sideload sequence. That isn't allowed (and you get a UVM_ERROR) if the
process was waiting for a grant. Handle things properly with a token
exchange mechanism. Since we've only got one process to control, we
could just use a flag, but this version seems a bit more robust.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>