[aes] Don't enter CLEAR state if triggers get zeroed during PRNG_UPDATE
The trigger register can be written by software while AES is busy.
Previously, it was possible for software to overwrite the triggers
for register clearing while the FSM was busy initiating the clearing.
As a result, the cipher core could get started without valid command
resulting in the cipher core FSM to enter the terminal error state
and signaling a fatal alert. With this commit, the main controller FSM
will simply abort and return to the idle state instead of starting
the cipher core FSM with an invalid command.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>