[sw/silicon_creator] Harden sigverify_keys
This change hardens sigverify_keys by adding redundant checks for the
life cycle state, key type, OTP read value, and a final redundant key
validity check before returning the key.

The order in which a key is checked is changed such that we first check
the life cycle state followed by the key type. This allows us to have a
single switch statement for the life cycle state which is easier to
harden.

Signed-off-by: Alphan Ulusoy <alphan@google.com>