[rom] Setup watchdog timer in assembly before entropy setup
This new code enables the watchdog timer right at the start of mask
ROM execution (before entropy setup and SRAM scrambling). This means
that unless the watchdog timer is 'petted' (i.e. the
aon_timer.WDOG_COUNT register is reset to 0 periodically) the device
will eventually reset itself.

The timeout value is currently the maximum possible (~6 hours assuming
a clock of 200KHz). This value should be reduced to something lower,
issue #10215 tracks this.

This change has been manually tested by setting the timer value to 0
and checking that the device is quickly reset.

Signed-off-by: Michael Munday <mike.munday@lowrisc.org>