[usbdev] Fix up packet responses for compliance
Add endpoint enable and stall registers, and correct responses for
unimplemented or disabled endpoints. Also correct behavior for SETUP
packets coming on non-control endpoints.

IN and OUT endpoints can be independently stalled, so they get
independent software control. Likewise, distinguish between endpoints
that are disabled (i.e. ignore transactions) and those that simply NAK.

Also add the spec-provided timeout for host responses. Detect the
start-of-packet (idle-to-K transition) and ensure it occurs within 18
bit times (with some fudging of numbers, since this should be at the
device's pins and is likely more liberal).

Clean up the register documentation for handling transaction handshakes.
Clean up the signal naming to reflect the correct terms, such as
referring to transactions instead of transfers.

Adjust the DIF and legacy usb code to use the new registers. Also delay
connecting the USB device to the bus until the default endpoint is set
up.

Tweak hello_usbdev to wait for the host to configure the USB device
before proceeding to the main loop.

Adjust the DPI code to better track the timing of responses to some
control requests, in addition to preventing the DPI instantiation from
locking itself out due to gating the function call that enables the
sense with a signal that depends on VBUS sense.

Tested hello_usbdev on FPGA and Verilator.

Signed-off-by: Alexander Williams <awill@google.com>