[tlul_adapter_sram] Ensure rdata is a specific value in error cases
This expands the sram adapter logic to return a specific value when a
TL-UL error is detected.

The value returned depends on whether the access was an instruction
fetch or not, and can be set via the DataWhenInstrError and
DataWhenError parameters.

It should be noted that the logic still returns correct bus integrity
for these values, since otherwise a regular TL-UL error would also
trigger a bus integrity error, which may not be expected.

Signed-off-by: Michael Schaffner <msf@opentitan.org>