[otbn,dv] Commit GPR changes when stalled
This only really makes a difference for BN.LID instructions like this:

    bn.lid x15++, 1184(x22)

Here, the RTL increments x15 on the first cycle. If the load triggered
an error because the DMEM integrity bits were malformed, all
side-effects of the instruction will have been discarded except for
the write to x15.

Until this change, the ISS committed the GPR change on the second
cycle. Of course, this meant that the aborting instruction would not
commit the GPR change at all, resulting in a mismatch between RTL and
ISS for the final state.

Note that this difference isn't actually observable in the design:
there's no bus access to register values and we will wipe the register
file between operations. But it *did* cause test failures because we
have a comparison of post-run dumps. Those should be fixed now.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>