[otbn,dv] Define a new class of trace entry for secure wipe writes
These appear from the RTL side when secure wipe is enabled and we will
see each register written with some pseudo-random value and then zero.
Once everything has been merged, we get something like the following:

    V
    > ACC: 0x00000000_00000000_00000000_00000000_00000000_00000000_00000000_00000000
    > ACC: 0xdf716993_b8b6de41_731c6c2c_ac85a1fc_9ea1db5c_1fc82f0d_ce1f8aeb_318d75a1
    > FLAGS0: {C: 0, M: 0, L: 1, Z: 0}
    > FLAGS1: {C: 0, M: 0, L: 1, Z: 0}
    > MOD: 0x00000000_00000000_00000000_00000000_00000000_00000000_00000000_00000000
    > MOD: 0x5bda6e0a_d54dc0cb_6d4b3ebf_a7cfe0b9_4191b56b_a739d816_ca87dfde_3bad8c9c
    > x02: 0x00000000
    > x02: 0x96ea3fb5
    > x03: 0x00000000
    > x03: 0xd4c44533
    > x04: 0x00000000
    > x04: 0xbea81560
    ...
    > x30: 0x00000000
    > x30: 0xebd16a69
    > x31: 0x00000000
    > x31: 0x2f472790

This commit doesn't teach the ISS to generate matching entries. Nor
does it handle the updates above properly: the "concatenate lines and
then sort" logic can't distinguish between

    x02 <= 0; x02 <= 0x1234

and

    x02 <= 0x1234; x02 <= 0

We'll fix that in the next commit.

Note that there used to be some logic in OtbnTraceChecker::OnIssTrace
to discard STALL entries with no changes. That was a bit of a hack,
which I added to avoid breaking things because of some spurious ISS
STALL lines after the operation had completed. This patch moves the
"discard STALL" logic to the Python instead which should make things a
bit easier to debug.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>