[opentitantool] Add non-SPI rescue protocol for bootstrapping
Ti50 firmware supports a serial protocol for flashing a replacement
image, which we would like opentitantool to be able to use.  In order
to add this protocol a few changes to generic code has been made:

1. Refactoring of bootstrap/mod.rs to remove assumption of SPI and
bootstrap pin being used for all protocols.

2. Minor modifications to io/uart.rs trait and implementations, to
ensure uniform behavior of read_timeout(), and to simplify callers use
of write().

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I2a612abb2f2de0b2f1f72cdc7bf94c0a51ec99de