[lib] Add `volatile` to launder*() to avoid miscompilations
In some cases, LLVM has shown it is able to deduplicate calls to
launder32() in a way inimical to hardening. Although this isn't
the "most correct" way to fix it, it is one that we know will work
for the time being until we need to revisit whether this volatile
is unnecessary.

The comments on the implementation have been extended with further
notes on volatile asm.

Signed-off-by: Miguel Young de la Sota <mcyoung@google.com>