[rom_ctrl,rtl] Fix a bug in the mux with invalid signals
This fixes a bug discovered by Prajwala with her signal corruption
tests. Suppose you corrupt the mubi4_t that appears here as sel_bus_i
some time after the ROM check has completed.

With the previous code, we sent chk_req_i (constant 1) to the ROM and
then reflected it back as bus_rvalid_o on the cycle after the glitch
went away. This, in turn, caused an assertion to go off in the TL
converter. In the real design, we might have even manufactured a stray
bus response. Oops!

This commit makes sure that we only make a request if sel_q is one of
the allowed mubi4_t values.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>