[rom_ctrl] Slightly beef up internal consistency checks
This was triggered by noticing that if we glitch the checker_done
signal then, while we would trigger an alert, we could also jump to
the Done state early, possibly hiding a difference in the ROM
contents.

With this change, we will jump to an invalid state. The alert signal
will still come out, but we'll never get to Done.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>