[rom_ctrl] Squash ROM bus responses on an internal error
There are a few awkward corner cases surrounding the mux, where an
injected fault can cause a response to come out as well as triggering
an alert. We've got some extra layers of safety here (e.g. ensuring
that the response data will have malformed integrity), but it's not
hard to explicitly squash things as well.

The result will be that the bus transaction hangs (and the chip will
eventually watchdog).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>