[rtl] Wire TL intg error from ROM adapter
Previously, this was connected from the ROM reg top, which tied off the
tl intg error signal to 0, assuming it will be computed by the next
stage.

The TL adapter was also instantiated to not check TL intg on incoming
command, (also fixed in this commit). The TL intg error is returned from
the ROM adapter instead of rom reg top.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>