[aes] De-assert idle after a new key becomes ready, flop idle_o
This commit modifies the generation of the idle bit in the status
register to prevent that writes to the main control, key and IV
registers following a write to the main control register or a key
register collide with the start of a PRNG reseed operation thereby
getting lost. This is required as with the key_touch_forces_reseed
feature, a write to the control register (if sideload is enabled) or
the write of the last key register may trigger a PRNG reseed operation.

In addition, the idle_o status bit for the clock manager is flopped to
match the corresponding bit in the status register. Previously, it was
one clock cycle off.

This is related to lowRISC/OpenTitan#10518.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>