[sw/silicon_creator] Move last_valid_entry out of page info struct
We currently have a field in boot_data_page_info_t that holds the last
valid boot data entry of an info page and  use two of these structs to
determine the active page. This change moves the last_valid_entry out of
this struct so that we can eliminiate a copy in `boot_data_read()` and
the second page info struct.

Signed-off-by: Alphan Ulusoy <alphan@google.com>