[kmac/rtl,DV] Alignment on swapping alerts order
This PR follows PR #10517 which swaps the order of fatal and recoverable
alerts. With that PR, now alert[0] is recoverable alert and alert[1] is
fatal alert.
With this change, we would like to update the alert_test register as
well, so now:
alert_test field[0] corresponds to recov_operation_err
alert_test field[1] corresponds to fatal_fault_err

Signed-off-by: Cindy Chen <chencindy@opentitan.org>