[chip dv] Indicate connectivity tests in testplan
The support for testplan tags has been added. This PR annotates the
testpoints in the testplan with the tag "conn" to better indicate the
connectivity tests. It also renames the testpoints to prefix "chip_conn"
as a standardized way to indicate all connectivity tests, which though
is redundant.

A minor change sneaked into this commit is renaming all exclusion files
in DV for signals that are covered in formal based connectivity tests
with the prefix "conn_" to more clearly represent them.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>