[sw/ottf] Add manifest to OTTF so mask ROM can (eventually) boot it
This adds a manifest field to any test images that are run via the
ottf_start.S assembly. This includes any tests that use the OTTF itself
(e.g., chip-level tests), and tests with a `main()` entrypoint (e.g.,
"Hello World" example programs).

Note, this just adds the manifest fields to OTTF test binaries, and sets
the default entry point to `_ottf_start`. It does not place the manifest
in the correct location such that it can be loaded and verified by the
mask ROM. That will come in a subsequent commit.

This addresses task in #10498.

Signed-off-by: Timothy Trippel <ttrippel@google.com>