[sw/ottf] move OTTF to flash address space
To demonstrate the flash translation capabilities of the Ibex core
wrapper, the OTTF was loaded at the virtual flash address space and the
Ibex wrapper IP's address translation features were used to map the
virtual flash address space to the physical address space.

Unfortunately, in order to run OTTF-launched tests with mask ROM (vs
test ROM), the OTTF manifest and entry point must be in the ROM_EXT slot
a (or b, though we load (currently) load slot a only for simplicity)
address space, wich is the beginning of the physical flash address space.

If the mask ROM evolves to make use of these translation features, then
the OTTF and test ROM can be updated to follow suit. This is captured in
issue #10712.

This partially addresses a task in #10498.

Signed-off-by: Timothy Trippel <ttrippel@google.com>