[aes] Re-enable synthesis optimizations for FWD/INV-only core instances
This commit moves the prim_buf on the cipher core operation signal to
the outside of the cipher core. Having this prim_buf inside the cipher
core prevents synthesis from performing area optimizations if only
forward or inverse operation is needed for a particular cipher core
instance (see e.g. CSRNG).

This is related to lowRISC/OpenTitan#10483.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>