[otbn,dv] Clear the "imem is invalidated" flag when loading program
This hasn't usually mattered because we reset the dut (and also the
ISS) after injecting IMEM errors to clear the locked state. However,
there was a race window where we injected the errors just after OTBN
had finished. In this case, we don't need to reset the DUT, but we do
have to clear the flag when the IMEM contents are replaced.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>