[flash_ctrl] ADD TEST FOR THE RMA FEATURE
Test the Flash RMA Functionality

Fash Partitions are Erased, Programmed and then Wiped by the RMA.  After the RMA has completed,
The test checks that SW Access to the FLASH is denied, and after a subsequent Reset, the Flash
is confirmed to be set to data other than that previously programmed.
Each run of the test performs 4 RMA cycles, each seperated by a Reset.

Signed-off-by: TIM EWINS <tim.ewins@ensilica.com>