[bazel] change name of mask ROM keys bazel build target
The bazel build target that contains the public keys to be included in
the mask ROM was named "test_keys". However, this target contains the
public keys. This commit renames the build target to be more specific,
i.e., the target name is now "test_public_keys".

This enables adding a filegroup target for each (test-only) private key,
which is needed to integrate image signing into bazel. This addresses a
task in #10864.

Signed-off-by: Timothy Trippel <ttrippel@google.com>