[usbdev] Clean up confusing uses of "differential"
Realign uses of the term "differential" to be from the perspective of
the usbdev IP. Wherever a differential mode is used, the IP is expected
to use differential signaling (i.e. with a P and N pair).

Previously, the term on RX would mean that an *external* differential
receiver would be used, delivering a *single-ended* data signal to the
IP. However, this top-level definition causes too much confusion.

Now, "differential RX mode" means the IP is itself the differential
receiver, and "single-ended RX mode" should be used with an external
differential receiver.

Similarly, for TX, "differential mode" means the data is output as a P/N
pair. "Single-ended mode" means data is output as the single-ended D and
SE0 signals.

Signed-off-by: Alexander Williams <awill@google.com>