[entropy_src/rtl] Remove gating on err_code status register
The REGWEN change has created conflicting conditions for testing the ERR_CODE register.
Two possible ways to resolve, either remove the redundant gating, or remove the
REGWEN keyword on the ERROR_CODE_TEST register.
Removing the enable gating makes it consistant with other bits in this register.
Removing the REGWEN keyword seems to be a system exposure so not doing that.

Signed-off-by: Mark Branstad <mark.branstad@wdc.com>