[ibex] Set ePMP reset values (ROM: LRX, MMIO: LRW, MMWP=1, RLB=1)
After reset ePMP will be configured as follows:

 | Entry | Address Space | Encoding | Permissions |
 |-------|---------------|----------|-------------|
 | 2     | ROM           | NAPOT    | LRX         |
 | 11    | MMIO          | TOR      | LRW         |

Machine mode whitelist policy (MMWP) and rule locking bypass (RLB)
will also be enabled.

This change also modifies the test ROM so that it sets up the ePMP
configuration such that read, write and execute accesses are
permitted anywhere in the address space to match the pre-existing
behavior.

Finally some ePMP configuration done by the mask ROM is redundant
and so has been removed. The mask ROM checks the ePMP configuration
is as expected and this code is not changed.

Fixes #7834.

Signed-off-by: Michael Munday <mike.munday@lowrisc.org>