[opentitantool] Proxy transport implementation
Add a new implementation of the Transport trait, which forwards every
method invocation via a network protocol to a separate session proxy
process.

This is to be used either with the session process being a software
emulator of sorts, or with the session process managing a USB
connection to a debugger device connected to hardware.

This is one of the steps outlined in issue #10889.  Implementation of
the session process is in PR #11054.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I6d4cd777a122e3017763704df6632f6f769991fd