[dv, jtag_agent] Add JTAG DTM RAL model, fixes
This commit enhances the jtag_agent in the following ways:
- Newly added JTAG DTM RAL model (hand-written) based off of
  riscv JTAG debug spec.
- Newly added JTAG RAL adapter to use JTAG sequencer
- jtag_item enhanced to use `i|dr_len == 0` to
  indicate what packet type it is.
- jtag_item enhanced to allow sending both IR and DR trans
  as a part of the same seq item, with an optimization switch
  to skip re-selecting the same IR that was selected before.
- Enhanced the jtag_driver to send both, IR and DR trans
  one after the other as a part of the same transaction,
  return response on `dout`.
- Removed `seq_lib/` directory to instead, have `jtag_seq_lib`
  since the sequences are tiny.
- Updates in external components to reflect these changes.
- Create the JTAG DTM RAL in `jtag_agent_cfg` and the adapter
  in `jtag_agent` directly, set the sequencer and adapter to
  `default_map` in the RAL. This enables the layering the agent
  up to JTAG DMI, and later, the chip CSRs via JTAG SBA DMI CSR.

The changes to the jtag agent forces any direct or indirect users
of it to change the tl_dw and tl_dbw settings, which are applied
to UVM_REG_DATA|BYTENABLE_WIDTHs respectively. This means a need
to update lc_ctrl and the full chip (and in subsequent commit, the
rv_dm_ testbenches. Ideally, we should have not overridden these
to begin with; hindsight is 20/20 I suppose. These changes hence
belong to this commit as well.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>