[opentitantool] Gracefully handle requests for unsupported interfaces
As we plan to introduce a session proxy process, it is not desirable
that the persistent session process could panic!() if a single command
line invocation asks for e.g. I2C, though the particular USB backend
does not support I2C at all.

We already have a TransportError value to be used if asking for e.g. a
UART instance which does not exist: InvalidInstance.  It would be sent
via the TCP protocol and presented by the command line invocation.

This change introduces a new InvalidInterface.  And modifies the
default implementation in the Transport trait to return that instead
of panicking, in case the transport does not support UART, I2C or
whatever kind of interface it is being asked about.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I7c0ff6e66a87ccd444be603be0075663d99cc280