[prim_dom_and_2share] Use prim_xor2 and prim_flop_en primitives
Previously, we were using inline synthesis attributes/comments which are
not portable to other tools. Instead, this commit switches to prim_xor2
and prim_flop_en primitives as used inside the DOM S-Box for AES to
place synthesis optimization barriers in a more portable manner.

This is related to lowRISC/OpenTitan#10823.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>