[util] Reorganize testutil autogen script for resuse
This reorganizes the core testutil autogen script logic so that it can
be reused by the `util/make_new_dif.py` script. This enable the
`util/make_new_dif.py` script to regenerate the autogenerated testutils
when the DIFs are regenerated, as is done in CI to check for stale DIFs.

This partially addresses a task in #9038.

Signed-off-by: Timothy Trippel <ttrippel@google.com>