[usbdev] Move pinflip and dp_o/dn_o to usb_fs_nb_pe
Handle the interface changes and alternate TX interface in logic, before
the final register. Prior to this change, the pinflip was sometimes
handled in clock muxes, on the I/O side of the first/final registers,
and the dp_o/dn_o interface was only done in behavioral logic. Move this
all inside the USB clock domain, and perform only the pin overrides with
clock muxes.

Add override for rx_enable and remove the redundant one for
tx_use_d_se0.

Signed-off-by: Alexander Williams <awill@google.com>