[jtag_agent, dv] Minor fixes, enhancements
This commit adds an intermediate `jtag_dtm_base_reg` class
from which all JTAG DTM registers are derived. The JTAG DR is
is updated and captured at the same time, so this intermediate
base class provides a mechanism to return the value to write
when performing reads. This is typically the mirrored value,
but individual JTAG DTM registers can choose something else.

Other changes are minor fixes and improvements.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>