[otbn,dv] Remove support for the model in the OTBN module
This was originally put in so that we could do system-level testing of
OTBN before the implementation was complete. Since then, it has been
used once or twice for performance tests where we needed to model the
Ibex side as well. But it's really rather difficult to maintain, and
leaves a bit of "DV-ish" code in the design folder.

If we need to do something like this in future, it shouldn't be too
hard to knock together another frontend to the simulator (like
stepped.sv) and to build a testbench in Python.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>