[entropy_src/dif] Disable flaky smoke test
entropy_src_smoketest performs an invalid check against a hard-coded
constant. It reads the data from entropy_src, but it does not control
the environment that generates that data. Consequently, various factors
can cause CI to break from this test spuriously failing.

Disable the test and downgrade the DIF to S0 until a replacement
strategy is implemented.

Signed-off-by: Alexander Williams <awill@google.com>