[otbn,rtl] Add cast to output of prim_lc_sync instance
This avoids a type warning with Xcelium. The point is that the lc_en_o
output of a prim_lc_sync with NumCopies = 1 has type lc_tx_t[1] which
isn't quite the same thing as an lc_tx_t.

The curly braces act like a reinterpret_cast and tell SystemVerilog to
just look at the bit patterns and stop worrying about it :-)

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>