[fpv/csr] Enable checking CSRs with regwen
This PR enables formal to check CSRs regs which could be gated by
regwen.
The regwen can be HRO or HRW, so we need to probe to see the regwen
value. Because DV and FPV testbench has a slightly different path, we
use `FPV_ON` macro to distinguish which path to use.

Signed-off-by: Cindy Chen <chencindy@opentitan.org>