[otbn,dv] Remove weak attributes for some more DPI functions
The cunning trick of using a weak symbol for a DPI export that may or
may not exist doesn't work on Xcelium. They were needed here because
we supported a use case for otbn_core_model where there was no RTL to
inspect. Fortunately, that use case is no longer supported (as of
commit 0d0fa1a09) so we can drop them entirely.

The other associated changes in the commit add some assertions to make
sure that the DesignScope parameter is a nonempty string.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>