[sw, base] Add a macro for creating addressable code labels
This macro is intended for writing chip-level tests that want to check
that some kind of exception-ey thing happened in a range of program
counters. It is a maximally compiler-portable implementation that won't
be optimization-hostile like GCC's label references (&&label), whose
support in LLVM is... not something we should rely on.

This macro is mostly made of caveats, but it has a narrow use
regardless.

Signed-off-by: Miguel Young de la Sota <mcyoung@google.com>