[dv] Make an implicit up-conversion explicit
Here, addr[i] is a uint8_t but C's confusing casting rules mean that
by the time it gets shifted left it has already been converted to an
int (which is good, because otherwise we'd be invoking undefined
behaviour when we shifted left by more than 8!). Make this widening
explicit.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>