[dv] Silence some trivial warnings in imported comline.inc
This code is copy-pasted from [1] and triggers some (completely
reasonable) compiler warnings.

Firstly, there's an unused variable. Secondly, we use scanf with a
format string implying a uint16_t and give the address of a uint64_t.
In practice, the latter will probably work on any little-endian
machine, but it's not a great idea.

Make the obvious changes to get everything compiling quietly.

[1] http://www.lightweightcrypto.org/implementations.php

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>