[dv] Silence some trivial warnings in imported present.inc
This code is copy-pasted from [1] and triggers some compiler warnings.

Firstly, there's an unused variable. Secondly, we've got some code of
the form

    A >> B - C

This generates a warning with most C/C++ compilers because people tend
to forget that it means A >> (B - C) rather than (A >> B) - C.

Drop the unused variable and add the parentheses to get everything
compiling quietly.

[1] http://www.lightweightcrypto.org/implementations.php

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>