[keymgr,dv] Fix some enum casting problems
The cast that we add in keymgr_scoreboard.sv to current_op_status is
to silence warnings from Xcelium where you have to be explicit about
converting from an integral type to an enum. Here, we know that the
conversion is safe because we've just checked that the value equals
OpDoneSuccess or OpDoneFail.

The other two changes also fix Xcelium warnings, but possibly change
the meaning of the code (for the better!). For example, in
keymgr_base_vseq, we are reading a 3-bit register. We were then
forcing it to be of type keymgr_ops_e, but that enum doesn't actually
have an item for the bit pattern 3'b111. Fortunately, we only ever use
the result to compare it with enum values so we can just store the
integral value instead, avoiding the cast. The only need for the
"cast_operation" variable is now to generate a debug message.

The same argument applies to current_design_state (which is used as an
output parameter to csr_rd).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>