[aes] Clear START trigger when ignoring it
The START trigger is ignored by the AES unit when the current
configuration is invalid (MODE = AES_NONE) or when performing automatic
operation (MANUAL_OPERATION = 0). When ignoring it, it's better to
clear this trigger bit back to 0 to avoid the unintentional operations
to start once a valid configuration is written or manual operation is
enabled.

This resolves lowRISC/OpenTitan#11166.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>