[opentitantool] Refactor TCP port search
When attempting to find an available TCP port to bind to, the current
code catches any error from the JsonSocketServer constructor and
handles it by calling the constructor again with the next candidate
port.

This change makes sure that unrelated errors to not get handled as
above.  The constructor is changed to take a ready TcpListener, and
the loop over available ports can then happen before the constructor
is called, and any Err() return from the constructor is now treated as
a fatal exception.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I765ba8897c2e7d20004056de92e260ef274144fd