[opentitantool] Bootstrapping efficiently via proxy
The proxy potocol allow issuing SPI commands to a USB debugger
connected on a remote machine running opentitansession.  This can be
useful for short interactions (e.g. testing TPM functionality), but
for bootstrapping by writing an entire image into flash, even the
slightest round-trip time will causing excruciatingly slow progress.
The device may even be sensitive to SPI transaction timing, causing
the bootstrapping to fail, if performed with significant latency.

Considering that opentitansession compiles with all the same library
functions for the SPI bootstrapping protocols as the "local"
opentitantool, it would be much more efficient to transfer the image
file to the proxy daemon, and let it break it down into individual SPI
transactions.

This CL does exactly that, by adding code to Bootstrap::update() for
detecting when the Transport implementation happens to be a proxy
stub, and in such case make use of a new "Transport::proxy_ops()"
method call to transfer the entire image file to the proxy daemon,
which in turn will invoke the same Bootstrap::update() which this time
will detect a non-proxy (USB) Transport implementation, and perform
SPI operations using it.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>

Change-Id: Icdc4eaa98368aaa3eaca0ce9edee971b29a4961b