[dv,jtag] Fix casting in jtag_riscv_monitor
This triggered some warnings in Xcelium (because we were assigning raw
bits to an enum type without an explicit cast).

Add explicit casting and an auxiliary "raw" variable to make it clear
which bits use the cast (now always inside a branch where we know the
value is good, or just used for debug prints).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>