[kmac] Make alert status stay
In issue #10760, @cindychip found that the alert status does not keep
its value as the status register is set to external register. As
`alert_operation` just asserts one cycle, the STATUS value becomes zero
at the next cycle.

In this commit, reg is defined to keep the value. It can be reset by
err_processed (not the alert but the STATUS value) for recoverable
error but the fatal error remains high until the IP is being reset.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>