[sw,tests] Verify flash_idle signaling to pwrmgr
For test: chip_sw_flash_idle_low_power.

Checks that when low power entry is enabled and a flash operation is in progress that the
low power entry is cancelled upon receiving the WFI instruction.
The watchdog timer barks to exit the WFI and a check is done to ensure the flash operation has completed.

Signed-off-by: Dave Williams <dave.williams@ensilica.com>