[otbn,dv] Tweak tracer to handle STALL + ERROR properly
A 2-cycle instruction normally generates a stall (S) trace entry and
then an execute (E) entry. If an error is injected (for example, with
the lc_escalate_en_i signal) that arrives on the 1st cycle, we want
the RTL to spit out the S entry and stop.

Before this change, the code that decided whether we were stalling
looked at state_d. Unfortunately, this doesn't quite work because when
we've got an error, state_d is OtbnStateLocked.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>