[otbn,dv] Drop a bogus assertion about OTBN RND interface
The logic behind this assertion was that the READING state only
happens when we're blocking on a CSR/WSR read of the RND register. The
controller will only drop its request for the data if something goes
wrong, which can't happen because it can't execute another instruction
to cause an error.

Of course, this is false: an error might come in from outside, either
from an illegal bus access or the lc_escalate_en_i signal.

Drop the assertion, since it just isn't true.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>