[otbn, rtl] Fix `idle_o` timing
`idle_o` will gate the clock if clock hints are setup to do so. Any done
interrupt raised must occur at or before the cycle `idle_o` changes to
ensure the clock gating doesn't erase the interrupt.

This adds a delay to the externally visible `idle_o` to ensure it
synchronises with changes to the status register and gives correct
interrupt timing.

Signed-off-by: Greg Chadwick <gac@lowrisc.org>