[opentitantool] Correctly detect bootstrap errors
Correction to the error handling in the bootstrapping protocol of
legacy Google Titan chips.  This change makes it such that if first 32
bytes of the SPI response are identical bytes, then it is considered
an error code (rather than random checksum mismatch).  Existing code
requires all 2048 bytes of the response to be identical, for it to be
considered an explicit error, and we have observed that opentitantool
will not always detect error codes such as BADKEY.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: Ied91475a5bb4f886eca79ece1e96718f3c5fbcf6