[spi_device] Block the passthrough with BUSY status
This commit addresses the issue #10785.

The BUSY @ SYS_CLK is set when SPI_DEVICE receives commands that satisfy:

- busy field in the CMD_INFO of the opcode is set
- CSb is de-asserted (SPI transaction completion)

When BUSY is set, SW may process the received command (assuming upload
field is set). After the command is processed by SW, SW may clear the
BUSY @SYS_CLK, which then will be propagated into SPI CLK (SCK) domain
at the next SPI transaction.

This commit is to turn off the passthrough interface if BUSY is set. So,
that the SW does not have to manually turn off the passthrough.

Keep in mind that this change cannot fully cover the corner cases:

- CSb is about to asserted (Beginning of the SPI transaction) at the
  same time BUSY is cleared by SW.

It is due to the CDC path of CSb to be used in SYS_CLK domain. It needs
to pass through 2FF synchronizer to be used in SYS_CLK safely. The delay
is 2 SYS_CLK. If BUSY falls in that window, the BUSY clearing may
unblock passthrough while SPI is active. The behavior of the IP is
undetermined.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>