[spi_device] Queue SW req of clr BUSY when active
Design Doc: https://docs.google.com/document/d/1wUIynMYVfVg9HmCL0q5-6r9BuN-XM0z--wGqU0bXRQ0

This commit adds a flag register to store the SW request of clearing the
BUSY bit in the STATUS register when the HW is not able to accept the
request due to the CDC issue.

The HW then checks the flag and clears the BUSY bit when the SPI is in
idle state.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>