[dv] Add ROM-only chip-level test example
A few chip-level tests (e.g., `chip_sw_flash_init`) must run and
complete entirely in the ROM stage. This means that the conventional
mechanism for running chip-level tests at the flash (ROM_EXT) stage is
not suitable for such tests.

This commit overcomes this by adding a mechanism to build a chip-level
test to be run in ROM using the meson build system. (Conveniently, the
`opentitan_functest` rule already provides such a mechanism using the
`test_in_rom` rule option.) Additionally, this commit updates the DV
testbench to skip loading of a flash image (either via bootstrap or via
the backdoor mem util mechanism) if no such flash image is provided at
runtime. Lastly, an example ROM-stage chip-level test is added to verify
the correctness of the added infrastructure, and to provide a reference
point for future chip-level test development.

Signed-off-by: Timothy Trippel <ttrippel@google.com>