[sba_access_utils_pkg] Improvemnts to SBA access APIs
This commit makes improvements to the SBA acess APIs, creating
a separate SBA access item class to capture the request and
the received response. Eventually, this item will be used by
a future SBA access monitor implementation.

The RV_DM SBA access vseq is also updated in accordance with the
changes made.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>