[ibex] Fix ePMP reset values
The ROM size was incorrectly encoded as 8KB rather than 32KB.
This occurred because the size encoding did not take into account
the two additional trailing bits in the reset value that are not
placed into the register.

Also the MMIO region should have the lock bit set so that it is
accessible immediately after reset.

Signed-off-by: Michael Munday <mike.munday@lowrisc.org>