[opentitantool] Conf files relative to current directory
Daemons should generally not keep a reference to the current directory as of when they were
started, doing so runs the risk of random references preventing the unmounting or other management
of filesystem volumes, due to some accident in the way a daemon was started.  For that reason, the
opentitansession daemon it started with the root directory ( / ) as its current directory.

However, the first business of the daemon process is to bind to a socket, and read configuration
files according to command line arguments.  Those arguments could include relative paths to some
files.  In order to support that, this CL delays switching to the root directory until the daemon
process has finished initializing, and at the same time when it drops its connection to the
terminal that was used to start it.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: Ia3216568b382ca04c12c7613822343052daefbe1