[doc] Update "getting started" diagram to show software first.
Because everyone needs to build software, the workflow seems easier if
the software is built first, and then the different workflows branch off
afterwards.

Also adjusts the FPGA flow to more accurately describe the steps
involved (bitstream must be loaded, and "loading" and "running" the
software are not really separate steps).

Signed-off-by: Jade Philipoom <jadep@google.com>