[otbn,dv] Fix off-by-one error in pick_bad_addr
In the known_mem class, "top_addr" is the size of the memory in
bytes (possibly badly named!), so we have to subtract one from the
last gap in the list to avoid possibly picking a value that's one too
big.

This wouldn't matter if this was actually a memory (it will still be a
bad address!) but it's a problem if we're using this function to pick
a bad CSR: numbers strictly bigger than 4095 aren't actually
encodeable.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>