[bazel] Add rules to build spiflash utility
The SPI Flash tool has multiple purposes. The primary is to interact
with Verilator simulations or FGPA instances of OpenTitan to reset the
chip and bootstrap SW into flash. The second is to split a device BIN
into SPI flash frames that can be used in DV sims to bootstrap the chip.

While the former is now handled by opentitantool, the latter is still
required to build a bootstrap flash image that can be use in DV sim.
This adds bazel rules to support building this utility, and its deps.

Signed-off-by: Timothy Trippel <ttrippel@google.com>