[otbn,rtl] Cache err_bits during a run
This is analogous to the previous commit, but this time we're caching
the contents of the err_bits register. Again, this won't make any
difference when secure wipe is disabled (because the error bits get
pulsed at the same time as done_core). But it's needed if secure wipe
is enabled (when done_core will be signalled some time after the error
is seen).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>