[rv_dm dv] Test drive compile-time seed
This commit test-drives the compile-time randomization method by
introducing a pre-processor macro `BUILD_SEED` which is set to
the DVSim generated random seed value. This macro is added to
`common_sim_cfg.hjson`, which will apply to all simulators.

For RV_DM, the pre-processor macro `RV_DM_JTAG_IDCODE` is removed
and in the env package, the `RV_DM_JTAG_IDCODE` parameter is set
to the `BUILD_SEED` preprocessor macro instead. On successive
simulation runs, the IDCODE value in the design has shown to be
set to random values, demonstrating that this method is working.
The IDCODE value returns 1 when `--build-seed 1` is passed on
the command line.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>