[aes] Fix clearing of data input registers
Previously, the write enable for the data input registers was set for
two clock cycles when clearing the registers. This caused the
data_in_qe_i signals used for status tracking to be high during the
first clock cycle when back in IDLE. As a result, the AES unit would
immediately start when running in automatic operation.

This is related to lowRISC/OpenTitan#11431.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>