opentitantool: Rustfmt fix
Rustfmt insisted on putting a semicolon after the `return` statement in the custom `bail!()`
macro.  However, having that semicolon caused the Rust compiler to emit a warning, when the macro
was used in an expression context (as in `match ... { _ => bail!(...), }`.  Adding an extra level
of curly brackets seems to satisfy both rustfmt and the Rust compiler.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I6735dc3ad40ac943347ddb8dc55cb6148f10e43f