[present] Rewrite TB to avoid non-freely licensed code
The stuff that was in crypto_dpi_present had been taken from
http://www.lightweightcrypto.org/implementations.php which may well be
intended for this sort of use, but isn't actually available under an
open source licence.

Fortunately, PRESENT is easy enough to implement, so this PR knocks
together some code to do so.

The testbench is also altered slightly, partly to allow us to write a
simpler C++ implementation, and partly to avoid some slightly odd
parameter names.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>