[spi_device] Delay read command statemachine one cycle
This commit pushes the read command state machin moving from
`MainAddress` to others by one cycle.

Issue has been reported in #11764. The address counter when the
statemachine enters `MainAddress` was one clock short. I assumed it was
correct value but based on the waveform, it has no time to latch the
last byte.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>