[bazel] fix protobuf repo dependency
The protobuf repo dependency was resolved via the `http_archive` rule.
However, it was downloading the protobuf repo from GitHub. Recently, the
name of the main branch of this repo changed from `master` to `main`
which caused a build failure when building from a clean cache.

This updates the protobuf repo dependency to be resolved via the
`git_repository` rule, which additionally allows us to anchor this
dependency to a specific commit.

Signed-off-by: Timothy Trippel <ttrippel@google.com>