[ipgen, rv_plic] Enable specific module_instance_name for multiple plic groups
Current ipgen for rv_plic only allow a single rv_plic group, and
generate the `ip_autogen` based on the template name. Adding a field in
the parameter `module_instance_name` to allow more plic groups to be
generated.

To support this, all the non-template files in `hw/ip_templates/rv_plic`
(except docs) are converted to mako template files.

It is a no-op for top_earlgrey, with the slight EOF newline change from
the generated files.

Signed-off-by: Cindy Liu <hcindyl@google.com>