[sw,tests] Enter RMA LC_STATE and check flash access and wipe
For tests:
chip_sw_flash_rma_unlocked
chip_sw_flash_creator_seed_wipe_on_rma

The system is initially started in Dev LC_STATE and data is written to all flash partitions.
The system is then set to enter RMA mode. After entry the flash is read again and checked
that the flash is accessible and that the contents have been wiped and differ from the
previously written values.

Signed-off-by: Dave Williams <dave.williams@ensilica.com>