[dvsim,xcelium] Fix sed commands to generate plusargs
I really messed up 0dc541c. I think that I'd checked the code with .*
and a non-empty string, then noticed that it also matched empty
strings. So I changed to .+, but forgot to check that .+ worked with
non-empty strings. It turns out that "+" for "one or more occurrences"
is only supported by sed with -E.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>