[otbn,rtl] Make sure to latch "non-core" error bits
I think I probably broke this in 75885e6. If you look at e.g.
otbn_core.sv, you'll see that we have err_bits_q and err_bits_d and we
latch errors until the next start signal.

I thought that I'd done the same thing in otbn.sv for errors that
don't come from the core, but I really hadn't! There, err_bits,
err_bits_d and err_bits_q are used to make sure that we only update
the (hwext) ERR_BITS register when a command finishes. That's fine for
signals coming out of the core because (as mentioned above) we latch
them there. But it isn't fine for signals generated at top-level.

Note that, unlike the other "error bits" registers, we don't bother
including a provision for clearing them. This is because all of the
signals are fatal so we should never be able to start another
operation (clearing the error bits) anyway.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>