[lib] Use quadratic probing as a source of entropy avalanches for hardnened mem*()
Quadratic probing is a buffer traversal algorithm typically used by
hashtables that avoids "clustering" of values as they are inserted. This
has the nice property that it maps the initial starting index into a
(non-cryptographically) pseudorandom permutation of the indices. This
means it has a very high randomization to implementation complexity
ratio. We can replace this with more complex permutation choices in the
future, but it's a much better starting point than the previous linear
access.

Signed-off-by: Miguel Young de la Sota <mcyoung@google.com>