[dv/chip] Add alert_ping_timeout test
This PR adds an aler_ping_timeout test in chip-level.
As the testplan describes, this test will configure the alert ping
timeout to a very small value - 2 cycles, and expect the alert ping to
always timeout. Then we will check if the interrupt is triggered
correctly, and if the local alert cause is correct.

Signed-off-by: Cindy Chen <chencindy@opentitan.org>