[chip dv] Fix coverage collection for chip XBAR tests
The coverage on all functional hierarchies were inadvertantly
being collected on chip level XBAR tests, which are supposed to
stub all TL hosts and devices. This was causing a larger coverge
numbers to be reported than intended. This commit fixes that.

We add ` xbar_build_mode` as a replacement for the `default` build
mode. This is done so, so that the chip level testbench which
reuses XBAR tests can append / override `xbar_build_mode` with
additional build-time settings. We use this method to specify
the right coverage collection hierarchy file for the XBAR tests.
This coverage collection hierarchy file is the same one as the one
use for the chip level `cover_reg_top` build mode - which restricts
the coverage collection only to TL interfaces on all hosts and
devices.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>