[entropy_src/rtl] Suppress re-assertion of interrupts
The entropy_data and observe FIFOs are cleared with the entropy_src
is disabled, but any interrupts saying they have data must be
cleared by FW.

As noted in issue #11951, there is also a delay between when the
IP is disabled and the FIFO is cleared.  This means that
even when FIFO interrupts are cleared after disabling, the interrupt
can reassert itself in this brief delay period.

This PR therefore suppresses these interrupt events from
retriggering when the IP is disabled.

Signed-off-by: Martin Lueker-Boden <martin.lueker-boden@wdc.com>