[kmac_app_agent] Support hosts that don't spot constant shares
This code was originally designed for keymgr, which has a check to
make sure that no share is all zeros or all ones. The ROM controller
doesn't have such a check, so we need to tweak the agent slightly.

Without this change, setting err_rsp_pct to 100 in the config might
still result in a KMAC response whose rsp_error field is zero.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>