[kmac] Always use randomness from PRNG when updating DOM multipliers
FPGA measurements suggest that the use of intermediate results for
re-masking the DOM multipliers causes first-order leakage. Now that we
have a full-width PRNG capable of generating 800 bits of pseudo-random
data per clock cycle, this commit changes the design to use the PRNG
output whenever the DOM multipliers are evaluated. If the multipliers
aren't evaluated, we keep forwarding intermediate results to the
randomness input to not create leakage upon changing the data inputs
only.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>