[bazel] fix bazel build due to flash CTRL DIF usage
There was some merge skew due to the use of some new flash_ctrl DIFs in
the tree, over the use of the old flash_ctrl.c driver. Note, it seems
there is a `keymgr_functest.c` and `flash_ctrl_test.c` that are using
both. This should be fixed in a future PR.

Signed-off-by: Timothy Trippel <ttrippel@google.com>