[rv_core_ibex] Move sim_sram injection point
This addresses #11965 in a slightly different way.
Instead of replacing sim_sram with scratch registers, this instead
moves the TLUL injection point to an exposed window inside the
rv_core_ibex register space.

This removes the need to use a "fake" physical address that mask
ROM does not open for access.

To ensure this mechanism can be re-used for gate level simulation,
prim buffers are added on both the request/respons channesl for tlul.

This should ensure the signals are not optimized away and can be directly
tapped in a netlist.

Signed-off-by: Timothy Chen <timothytim@google.com>