[flash_ctrl] Separate invalid and disabled states.
- fixes #11825

Part of the D2S design fixes was to unify the disabled behavior across
more of the flash controller state machines. However this created a
problem for RMA entry.

When RMA entry is done, the RMA fsm also requests flash operationst to be
disabled. Previosuly, this caused the RMA fsm itself to transition into
an invalid state.  This caused 2 problems:

1. At the end of RMA, a fatal alert is generated. While this is
   not bad behavior in any meaningful way, it is very inconsistent with alert
   defintion because no fatal event has occurred. This leads to unnecessary
   DV work.

2. Since the RMA fsm transitions out of its ack state, it means the ack
   indication sent to life cycle controller will be held for only 1 cycle.
   This is most definitely not enough for CDC and means RMA cannot be completed.

This commit separates the idea of "disabled" and "invalid" for most of the fsms.
Disabled simply means functionally disabled from future operations, while
invalid is both disabled AND alert generation.

At the end of RMA transition, the device is simply "disabled" and not "invalid".
Invalid then can only be caused by an explicit error, or the FSMs going
completely to an undefined state.

This makes the overall definition more consistent.

Signed-off-by: Timothy Chen <timothytim@google.com>