[sw] Fix function prototypes
Several functions without parameters were declared as `foo()` instead of
`foo(void)`. The former is considered a function without a prototype, not
a function without parameters. Newer versions of Clang implement a
stricter check of this condition, and the new warnings were breaking the
build. Fix this by adding the missing `void` in several places.

Signed-off-by: Luís Marques <luismarques@lowrisc.org>