[otbn,rtl] Shuffle timing of some internal signals
The point of this is to allow us to flop the idle_o signal (through a
prim_mubi4_sender) and the intr_done_o signal (by adding a flop to the
prim_intr_hw instance).

It turns out that we can do this without changing the external timing
at all. The following signals move forward a cycle:

    locked                           (now called locking)
    otbn_dmem_scramble_key_req_busy
    otbn_imem_scramble_key_req_busy
    idle
    done

and we rename the is_not_running signals to have an explicit _d and
_q. Note that these don't quite line up with the is_not_running and
is_not_running_r that we had before: is_not_running_q is the same
timing as is_not_running.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>