[ci] Modify Bazel timeouts to reduce test failures
Increase the maximum allowed runtime of the Bazel job to allow more time
to populate the cache in the event of cache misses. This should be
relatively infrequent but can reduce unnecessary Bazel job failures.

Reduce the remote cache timeout to reduce the amount of time Bazel
spends waiting on the cache in the event of a flaky connection to
the GCP bucket.

Signed-off-by: Miles Dai <milesdai@google.com>