[opentitantool] Tweak to Ti50 rescue bootstrapping
The Ti50 rescue protocol transfers on the RW part of the image file provided (as it is the RO code
running on the chip that is receiving and doing the flashing).  A change in the image format
(encrypted cryptolib), causes the heuristic in the existing code to misinterpret where in the
image the RW section started.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: Ifac272f231535ecfe5f19c8f6b7423fdbaabeec7