[bazel] Add sha256s for http_archives
For archives with stable files, the sha256sum is computed locally and
added.
Archives with unstable files use a dynamic_archive repository rule and
check the sha256sum from the contents.
A singular sha256 allows bazel to draw from the cache and to perform
checking even when the archive file has a varying timestamp.

Dynamic archives now also emit correct prompts regarding how to add a
sha256 to make a canonical reproduceable form.

Tested:
For the new rule I fetched with an incorrect sum and it failed
I fetched with no sum and it succeeded and printed to canonical
reproduceable checksum
I set the checksum to the canonical reproduceable form and it passes.
This wont pass CI until I get the same checksum on my workstation as CI
so that offers some validation that I'm not leaking system specifics
into the check.

Signed-off-by: Drew Macrae <drewmacrae@google.com>