[flash_ctrl] DEBUG and Enable the OTP_MODEL
For some reason, the OTP_MODEL in the Testbench had been commented out.  This
left the acknowledge signals floating, and hence the otp interface locked up.
I have reviewed the otp_model, and have re-enabled it.  This is now working
with the test flash_ctrl_hw_sec_otp, and is and is safe to be present in all
other tests.  The otp_model is accessed through a vif between TB and the base
sequence.  The model provides random otp keys on demand.

Signed-off-by: TIM EWINS <tim.ewins@ensilica.com>