[spi_device] Update flip/ watermark ONLY when active
This commit revises the readbuffer module behavior. The module updates
and reports the flip/ watermark events only when active.
The active signal represents the exact time the buffer needs to be
updated. The signal comes from the upper module `readcmd`. It is a pulse
signal that represents the last beat of a byte. (e.g the second beat in
Quad mode).

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>