[spi_device] Watermark and Flip handling
This commit revises the watermark logic to handle the corner case
correctly. The case is when `watermark` and `flip` both events occur at
the same cycle.

If both events happen at the same cycle, it means host system issues a
read request to the next buffer above the threshold.

For example, the threshold is configured to 0x83 (131 Byte) and the
current buffer is {1kB : 2kB}. Then when the host system issues read
request to 0x486 (1,158 byte), it triggers `flip` event and `watermark`
events.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>