[bazel] support pre-initialized bitstream cache
If the bitstream cache has not been pre-initialized before running bazel
with the `BITSTREAM=--offline` environment variable present, then the
bazel bitstream fetching script will error out, as it won't be able to
locate a single bitstream.

This commit updates the bitstream repository rule to use a
pre-initialized cache if the `BAZEL_BITSTREAMS_CACHE` environment
variable is set.

Additionally, this commit updates the bazel airgapped prep script to copy
the pre-initialized bitstream cache to the `bazel-airgapped` directory
that is designed to be pre-installed on an airgapped machine.

This fixes #12455.

Signed-off-by: Timothy Trippel <ttrippel@google.com>