[otbn] Harden instruction fetch response address register
This commit prevents glitching of the register that stores the
instruction fetch response address.  It does so by instantiating a
`prim_count`, of which the *counter* functionality is not used, but the
redundant register, cross-checking of register value, error raising, and
the DV integration of `prim_count` are used.  The synthesizer should
optimize the tied-off counter logic away.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>