[bazel] Add enough verilator tags to reliably filter it out of the build
The test_suite contains verilator tests so we really want to filter it out
when we exclude verilator tests from targets to be built. If we also tag the
//hw:verilator targets that have a verilated model, we should be able to
build the hw directory and the whole software directory without relying
on clever tricks. (Let's keep the clever tricks for a bit)

Signed-off-by: Drew Macrae <drewmacrae@google.com>