[clkmgr] Handle measurement control on calibration lost
- addresses #12523

The basic problem described in the issue measurement controls survive
low power entry/exit.  However, the clock calibration settintgs do not
(the ast registers are reset).

This means upon low power exit, assuming measurements were configured
prior, there is an immediate mismatch between count and clock.

To address this, the clkmgr now observes when the clocks lose its
calibrated state by observing `ast_init_done`.  When this is seen,
clkmgr automatically disables measurements and allows software to
reconfigure.

While this solves the existing issue, it does mean that if an attacker
were able to fake an unlock, the measure controls would turn off.
However, there is no good way to handle that situation in this boundary
since that attack can look no different from a low power entry.

It will be up to the system to place defenses elsewhere to check that a
calibration lost is intended behavior.

Signed-off-by: Timothy Chen <timothytim@google.com>