fix(spi_device): Not push to SRAM when !FwMode
This commit fixes the wrong behavior of fwm_rxf_ctrl module. When the
spi mode is not FwMode, fwm_rxf_ctrl, fwm_txf_ctrl should not operate.
The fwm_txf_ctrl only fetches data from DPSRAM when write pointer has
been updated. If it is not FwMode, it is assumed that SW won't touch
write pointer of TXF.

However, the RXF runs based on the read async fifo valid signal. The
FIFO has been updated regardless of the spi mode. This commit adds
prevention logic in front of the async FIFO and also inside the FSMs.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>