[otbn,rtl] Avoid a 1-cycle glitch in locking_o
The locking_o signal used to be computed as follows:

  assign locking_o = (state_d == OtbnStateLocked) & ~secure_wipe_running_i;

This used to use state_q, but 48e4bb6 changed to state_d to move it a
cycle earlier. Of course, this doesn't work properly because you end
up with a single cycle "glitch" where locking_o is asserted before the
secure wipe starts.

The problem is that you can't just assert secure_wipe_running_o in the
start/stop controller a cycle earlier because you end up with a
combinational loop (I tried!). This commit switches the start/stop
controller to a sort of req/ack interface where the controller asks to
start a secure wipe and the start/stop controller responds once it's
done.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>