[bazel,dvsim] remove gen-binaries.py's reliance on meson_init.sh
Meson will soon be removed from our project and replaced entirely with
bazel (#12449). This updates the the OTBN `gen-binaries.py` script to
not rely on the presence of the `.env` file that was produced by the
`meson_init.sh` script to provide locations to the RV32 toolchain.

Instead, the `gen-binaries.py` script now uses environment variables to
get the locations of the RV32 toolchain tools, which are populated via
queries to bazel.

This fixes #12447.

Signed-off-by: Timothy Trippel <ttrippel@google.com>