[sca] Disable peripheral clocks during SCA capture
To further reduce unwanted noise during SCA captures, it's better to
disable the peripheral clocks in addition to the clocks of the bigger modules
like (HMAC, OTBN, KMAC, AES).

Special care is required for the IO_DIV4_PERI clock as the UART and GPIO
modules required for communication with the scope are using this clock.
Therefore, this clock can only be disabled during the actual capture and
has to be re-enabled afterwards to resume communication.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>