[dv] Cleanups in TL error vseq's tl_instr_type_err task
This was prompted by an Xcelium warning caused by the fact instr_type
was of type bit[3:0] and was passed to a port expecting a mubi4_t.

Because it seemed a bit tidier, I've refactored things a bit so that
the randcase just picks values that will be passed to a tl_access task
that runs afterwards. We're also now randomising data properly in the
second case, which I think we'd missed before.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>