[otbn,dv] Silence an Xcelium warning from otbn_core_model.sv
Since this code is inside a process, local variables default to static
lifetime. Xcelium spits out a warning if you don't say that
explicitly (because not noticing static lifetimes is an excellent way
to introduce silly bugs!). However, there's no strong reason not to
just inline the constant so let's do that.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>