[kmac,dv] Fix types in a digestpp_dpi helper function
The comparison between an int and a uint64_t was causing
a (reasonable!) compiler warning. Fix that, and also remove an unused
variable (arr_ptr was written but not read).

Once that's done, there's no point in the short-circuit when array_len
is zero, so we can remove that too.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>