[otbn,dv] Don't connect up "idle_o" output in TB
The checking for this signal is handled by the otbn_idle_checker
instance which is bound in to the otbn module. The connections that
we'd wired up here didn't actually do anything, and in fact were
wrong! This signal turned into a mubi4_t back in March (6a18e34) and
the only result was an Xcelium warning about mismatching port widths.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>