Changes to environment and rand_output_vseq
pwm_scoreboard.sv - changed item generation to ensure items are not generated
for duty cycles of 0% or 100%. Corrected the phase count and duty cycle
calculations and tidied up code.

rand_output_vseq.sv - Changed the duration cycle constraint to reflect changes
within the environment package, added phase delay randomisation and corrected
order of channel enable and inversion.

pwm_env_pkg.sv - A maximum clock divider paramter was introduced to constrain
the clock divider value to a maximum 4bit value to allow for practical testing
and simulation. Greater values of CLK_DIV can be tested within the stress test.

Made changes to the number of cycles per sequence, removing
the maximum and minimum values in favour of a different approach; the worst
case scenario for a 4bit CLK_DIV value was calculated and chosen as the default
sequence runtime. This will allow for multiples of this value to be used in
stress tests for different values of CLK_DIV.

Signed-off-by: Jason Hoddinett <jason.hoddinett.ensilica@opentitan.org>