[dv] Update prim_sparse_fsm_flop_if to force at the flop
To solve the FSM coverage issue, we deposit invalid value to the FSM
at these 2 places.
1. the __q at the same hierarchy of the state machine
2. state_o in the prim_sparse_fsm_flop
But `state_o` isn't the actual flop, after depositing the value, it
won't be changed unless `state_d` is updated, while __q will be changed
back to __d in the next cycle.
changed to desposit value on the flop rather than `state_o`

Signed-off-by: Weicai Yang <weicai@google.com>