[otbn,rtl] Tweak how we check integrity for ISPRs
This was triggered by a slight bug in the previous
implementation (where a CSRRS caused ispr_rdata_into_ispr_bignum_wdata
to be high, meaning that a CSRRS on RND_PREFETCH caused an error).

However, the previous design also doesn't really match the rest of
OTBN. In particular, an instruction like BN.WSRR reads ISPR data. It
writes that data (together with the integrity bits) straight to the
bignum register file. With the previous design, it didn't check
whether these bits were valid or not. This doesn't match the rest of
OTBN, where instructions like BN.MOV, BN.MOVR, BN.LID and BN.SID pass
integrity bits through unchanged *but* also check them for
correctness.

This commit tweaks things to work the same, and also writes a big
comment to explain what's going on (because it's a little fiddly to
understand all the corner cases).

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>