[otbn,rtl] Fix overflow in stack_top_idx
We use next_stack_top_idx to derive next_top_valid_o, which says
whether the stack has data in it. This check is based on seeing
whether the index just above the top of the stack is nonzero or not.
We were using StackDepthW bits and got an overflow when the stack
filled up.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>