[sw,tests] Test sysrst_ctrl combo detect reset and wakeup
For tests:
chip_sw_sysrst_ctrl_gsc_reset
chip_sw_sysrst_ctrl_sleep_gsc_wakeup
chip_sw_sysrst_ctrl_sleep_gsc_reset

Tests that a specified input pin combination can be detected by the sysrst_ctrl and reset the system.
Checks that the reset also asserts ec_rst_l when this is set as an action for a combo detect.
Checks that when a reset condition is entered that flash_wp is also asserted.
Tests that ULP wakeup in sysrst_ctrl can be triggered with a pin input to wake the system from sleep.
Tests that when the system is in deep sleep that an input pin combination can be detected and reset the
system.

Signed-off-by: Dave Williams <dave.williams@ensilica.com>