fix(ottf): Make SWI asynchronous ISR
https://github.com/lowRISC/opentitan/issues/12787

SWI in `ottf_isrs.S` was synchronous, where the logic changes MEPC by
adding the instruction size.

However, SWI in Opentitan (and according to Spec) is asynchronous. In
OpenTitan, after setting MSIP (Memory Mapped CSR in RV_PLIC), the
software interrupt occurs two cycles later due to the fabric latency and
the register update in RV_PLIC.

So, `handler_software_isr` is revised to store MEPC directly to the
stack not manipulating it.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>