[otbn, dv] Fixes regression issue in otbn_zero_state_err_urnd
The following issues were present in otbn_zero_state_err_urnd_vseq if
the zero state error was injected before the instruction fetch had
started:
1. The values deposited on wakeup_iss and xoshiro_q never changed after
the execution of release statement as there were no active drivers on
them.
2. reset_if_locked was executing too soon and not waiting for the status
to go to locked and hence the dut was not getting reset.
3. The parent vseq called run_otbn task while the reset was asserted if
reset was asserted immediately after the status went to locked.

This commit fixes all the three issues.

Signed-off-by: Prajwala Puttappa <prajwalaputtappa@lowrisc.org>