[otbn,rtl] Refactor ISPR read muxing
ISPR read data is muxed out in two stages in the bignum ALU. The first
stage muxes between all of the ISPRs that don't have integrity bits. The
result goes through an integrity encoder and is fed into the second
stage. The second stage chooses the final read data choosing between the
first stage with the calculated intergrity and the ISPRs that have
integrity bits.

This refactoring makes this structure more explicit and should result in
area savings as the previous RTL fed the same signal into multiple
inputs of the second stage mux. As there is an optimization barrier in
the mux these would not have been optimized down to a single input.

Signed-off-by: Greg Chadwick <gac@lowrisc.org>