[fpv/rstmgr] Fix rstmgr_sec_cm errors
This PR fixes the cex from rstmgr_sec_cm. In sec_cm test, we cut off the
logic for `state_o` (connects to state_q), and expects to trigger alert
is state_o is not one of the declared enum type.
This might cause issue for the following FSM code:
 unique case (state_q):
   FsmError: begin
     fsm_err_o = 1'b1;
   end
   default: begin
     state
If we continously driving non enum values to state_o, the `FsmError`
state will never be reached. To avoid this issue, we added an assumption
for this.

Signed-off-by: Cindy Chen <chencindy@opentitan.org>