[otbn,dv] Remove stray commit() call when starting ISS
Looking at the commit log, I think this dates back to when we passed
the start address in a weird way. The call to commit() ensured that
this got "registered" successfully. Anyway, we got rid of all of that
a long time ago, so let's just drop the confusing-looking call to
commit() as well.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>