chore(cdc): Prim Reg CDC waivers
the `u_core_tgl` is to catch the edge of the slower clock. The signal is
used to update the value to the faster clock domain.

The reconvergence error complains the signal combines with the src_ack.
However, src_ack comes from the slower clock domain and related to the
src_req(busy_q). Update signal is other one. If both happens, it is
still OK. If both happens at the different clock phase (1clk apart),
still OK as the intention is to update the registers in slower clock to
the faster clock.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>