[otbn,rtl] Squash prefetch mismatch error when controller stops
Without this squashing behaviour, there's a problem if the controller
stops on an instruction that doesn't look like a jump/branch/etc. This
stalls the instruction stream (so insn_fetch_req_addr_i stops
changing). In this case, the invalid instruction didn't look like a
jump or similar, so the prefetcher has already been incremented by 4.
We then generate an error (insn_addr_err_o) when we look at the
results.

Hopefully, any situation like this would be caught by the existing
testbench (by looking at ERR_BITS). But we definitely catch situations
where the initial error is only supposed to send out a recoverable
alert. In this case, the testbench notices that an unexpected fatal
alert came out.

Signed-off-by: Rupert Swarbrick <rswarbrick@lowrisc.org>