[ci] Add `notest_keep_going` flag to cw310 tests
The CW310 board continues to see UART failures frequently. This causes
all tests to fail with the UART timing out on opentitantool. This
continues until the entire job times out at 30 minutes which leads to a
long queue of jobs on the FPGA runner. With this flag, we fail
immediately to prevent waiting on a test will ultimately fail anyway.

Signed-off-by: Miles Dai <milesdai@google.com>