[sw/silicon_creator] Update functests that rely on reset info
Since the SPI flash based bootstrap implementation in opentitantool ends
with a software requested reset, both POR and SW_REQ bits of RESET_INFO
will be set when a functest runs. This commit updates functional tests
that expect rstmgr's reset info to be only POR at the start of the test.

Signed-off-by: Alphan Ulusoy <alphan@google.com>