[dvsim] issue `bazel clean` before `bazel build`
The ralgen.py script is using relative paths to find the topgen.py
script. This caused issues when migrating dvsim.py to use bazel over
meson, specifically two consecutive invocations of dvsim.py without a
bazel clean in between seemed to cause the topgen.py not to be
found.

PR #13071 tried to fix this, but had to be reverted in #13088 due to
issues with the dvsim.py scratch directory being outside the REPO_TOP.
Instead this updates the sim.mk file to always issue a `bazel clean`
before the `bazel build` which should also fix the issue #13071 tried to
fix.

Signed-off-by: Timothy Trippel <ttrippel@google.com>