[crypto] Add Ed25519 point encode/decode operations and tests.
An important decision here is how, in the decode operation, to determine
if an encoded point is valid. Details are in the comment at the top of
ed25519.s, but I have chosen here to use the ZIP15 validation criteria
to maximize predictability and compatibility.

Signed-off-by: Jade Philipoom <jadep@google.com>
(cherry picked from commit afeded2549e0efca41d819d8883e87b1e1c20c41)