[mask_rom] Add ROM_EXT measurement to the keymgr.
This commit adds an OTP config option (`ROM_KEYMGR_ROM_EXT_MEAS_EN`) to
the OTP owner partition. When this fuse is set to `kHardenedBoolTrue`,
the mask_rom will use the digest of the ROM_EXT to configure the key
manager attestation binding registers.

The ROM_EXT digest is also copied over to the static_critical data
section in case `ROM_KEYMGR_ROM_EXT_MEAS_EN` is set to
`kHardenedBoolFalse`, requiring the ROM_EXT to combine this value with
the next boot stage measurement.

See lowRISC/opentitan#7652 for more details.

Signed-off-by: Miguel Osorio <miguelosorio@google.com>