[otbn,rtl] Remove reset from FF-based regfiles
This removes the reset from the flip-flop based general-purpose
registers (GPRs and wide data registers (WDRs) as well as the `ACC` and
the `MOD` register.  This improves security:  If a register holds a
random mask to be applied to some other register (remember that the
masking is implemented in SW here) and an attacker manages to reset that
register while OTBN is running, the masking is effectively switched off.

Co-authored-by: Andreas Kurth <adk@lowrisc.org>
Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>