[otbn,rtl] Assert wdata known only if writing
Prior to this commit, the `DmemWdataOKnown_A` assertion would trigger
also on read requests (i.e., when `dmem_req_o` is set but `dmem_write_o`
is not).  For read requests, however, no data is written and thus the
write data value does not have to be known.  This caused spurious
failures of this assertion.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>