[opentitanlib] Add I2C for Ti50Emulator transport
The implementation of the I2C bus for the Ti50Emulator is based on the following assumptions.
Connections between DUT and the Host are made directly (point-to-point).
There are no other devices on the bus, only DUT and Host are connected.
In order to perform the tests, the connections do not need to be simulated at the level of transmitted bits.
The implementation does not simulate the ACK / NACK mechanism.

The emulated I2C BUS is implemented as a Unix Stream socket with a request/respons
based protocol to describe read and write transactions.

Signed-off-by: Michał Mazurek <maz@semihalf.com>