[otbn,dv] Fix reset during EDN request
We were having an assertion saying we shouldn't recieve EDN
words without a request already going. That is not necessarily true
if we are in the middle of a EDN request and resetting OTBN.
Request drops but we still end up with a word to process.

Now we are ignoring the incoming EDN word if we have not initialize
_acc properly.

Also, another bugix for forget function in RNDReq. Now it also clears
the request flag.

Signed-off-by: Canberk Topal <ctopal@lowrisc.org>