[opentitanlib] Add capability to clear UART RX buffer after bootstrap
This commit modifies the bootstrap implementation to clear the UART RX
buffer by default while holding the target in reset. This commit also
adds a new bootstrap command option `--no-reset-and-clear-rx` to make it
possible to override this behavior if needed, e.g. to observe the complete
output not just what is received after bootstrap is completed.

Signed-off-by: Alphan Ulusoy <alphan@google.com>