[util] Use correct working dir in generate_compilation_db.py
Specifically, this PR updates the script to emit bazel-opentitan/ as the
working directory for each compile command.

I used `bazel build -s` to print out subcommands when generating a test
target, and noticed that it was changing directory before each of the
calls to gcc, whereas the generated compile_commands.json just points to
the repo's root directory. Helpfully, Bazel already makes a symlink to
the correct working directory: bazel-opentitan/.

Signed-off-by: Dan McArdle <dmcardle@google.com>