[usbdev] Fix OUT ZLP handling under FIFO errors
Previously, zero-length packets could elicit an ACK response even when
the AV FIFO is empty or the RX FIFO is full. In the OUT transaction
state machine, the state of the FIFOs was only checked when data would
be written to a buffer. However, in usbdev_usbif, the transaction would
be rejected, so the host would incorrectly think its packet reached the
device.

Fix this up by having the state machine also check FIFO states upon
reaching the decision point for a handshake phase. This will be the
final gate that will reject ZLPs that can't get an entry for the RX
FIFO.

Signed-off-by: Alexander Williams <awill@google.com>