[opentitantool] Fallback to empty transport
In cases where an invalid `--interface` is specified it is better to
attempt to proceed with an empty transport than it is to fail
immediately. For instance, users may have a
`~/.config/opentitantool/config` file that specifies a particular
interface that may not always be available. Right now this will cause
all opentitantool operations to fail, even in cases where opentitantool
doesn't depend on any transport.

Signed-off-by: Jon Flatley <jflat@google.com>