[dv/cip_mubi_cov] Add a wrapper to support binding an mubi array
In alert_handler, lpg groups are an array of mubi inputs. To collect
coverage on mubi types, we usually bind the mubi input with
cip_mubi_cov.if. But binding an interface won't work for genvar for
loops. To avoid hardcoding 22 signals, I added a cip_mubi_cov_wrapper
module that uses a parameter to input number of mubi signals.

Signed-off-by: Cindy Chen <chencindy@opentitan.org>