doc(hmac): Remove ambiguous description in MSG_FIFO
This commit removes the "strict order" part in the MSG_FIFO section. As
HMAC register space aligns with the MSG_FIFO window, it is likely that
the MSG_FIFO space is defined as 'device region'. With this
configuration, only one outstanding transaction the core issues to the
MSG_FIFO window.

As described above, the strict order is likely to be maintained in both
case (fixed word address vs MSG_FIFO window). The only benefit is that
cores can use store multiple commands.

Related issue #13497

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>