[usb_testutils] Avoid buffer double-free
The control endpoint code had a path that would lead to a double return
of the buffer. Check that the buffer is still active before calling the
DIF to return it.

This case is due to the buffer read function automatically returning the
buffer underneath when all data has been read.

Signed-off-by: Alexander Williams <awill@google.com>