[dvsim] use Bazel labels for SW images
This commit enhances the interactions between dvsim.py and Bazel to
partially address #13430.

First, this updates all dvsim configuration files to
use Bazel labels to reference SW images, as opposed to a sudo-path like
string to the Bazel target.

Second, this fixes a slight error in the building of the `bazel_opts`
string.

Signed-off-by: Timothy Trippel <ttrippel@google.com>