[crypto, test] Test timing properties of AES-GCM decryption.
Add a test that checks if AES-GCM decryption is constant-time relative
to different invalid tags. This helps ensure that an attacker cannot get
information about how much of the tag they've guessed correctly.

As part of this change, refactor AES-GCM testing so that utilities can
be imported by multiple files.

Signed-off-by: Jade Philipoom <jadep@google.com>