[otbn,dv] Add regression for CI
This commit creates a regression test for OTBN that contains all tests
that are supposed to pass according to the current development stage.

This could be run in CI instead of the `smoke` test.  The problem with
only running the `smoke` test on a PR is that tests that are supposed to
pass are not tested for each PR.  Thus, a PR can break `master` and it
can only be noticed in the next nightly run of regressions.

As all tests with their configured reseed multiplier probably take too
long to run in per-PR CI checks, the reseed multiplier can be reduced
through `dvsim`.  For reference, with `--reseed-multiplier 0.1`, this
commit takes 8 minutes to build and run in VCS on my laptop with 16
threads.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>