[dv, tl_agent] Invalidate a_channel before reset
At the chip level, in stub_cpu mode, we attach our TL agent,
which injects TL traffic over the CPU's data channel port.
The reset to this TL agent is the same reset hooked up to the
CPU, which de-asserts much later in the chip's system level
reset sequence. Until the first reset asserts, the TL host driver
does not set `a_valid` to 0, which means until the CPU's reset is
released, `a_valid` exhibits a value of X. This X ends up propagating
to downstream devices in the XBAR, whose resets get released sooner.

To prevent this X-prop, we invalidate the a_channel (and set the
`a_valid` to 0) at t=0, rather than waiting for the reset attached
to the TL interface to be released.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>