[dv/chip] remove forced lc_dft_en value
This PR removes the tb.sv file that force lc_dft_en to On.
This is not needed because the loaded image is "rma" which enables the
lc_dft_en signal. And the test waits until otp_init is done, so we won't
drive any tlul transactions when `lc_dft_en` is Off.

Signed-off-by: Cindy Chen <chencindy@opentitan.org>