[dv, xccelium] Disable assertions on $finish
This commit kills assertions in specific modules that
do not predictably complete when $finish is called. One of the
modules is `prim_reg_cdc`, which is used all over the design.
The CDC assertion starts every few clock cycles, and its
implication is satisfied a couple of clock cycles later.
Unfortunately, it is not possible to predictably call $finish
(i.e. end the sim) after the implication is satisfied. So
we just kill the assertions in these blocks in a `final`
block. This is ok, because these assertions fire several times
over during the course of the simulation, and complete successfully.
Its only the last one that causes a failure to be signaled
and cause the test to fail) AFTER the simulation is complete.

Strangely, this failure enhibited by Xcelium, not VCS.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>