[dv, waves] Improve wave dumping
A minor non-functional change to waves.tcl improves the way
wavedumpScope method is used to dump specific hierarchies
rather than dump the tb top by default.

The code is rearranged a little, and several comments are added
to make the usage clearer. the `fid` argument of `wavedumpScope
proc is removed in favor of using the value internally by
referencing a global variable instead of passing it as an arg.

Wavedumping has been tested with vcs:fsdb,vpd,vcd and
xcelium:shm,fsdb and everything seeems to be working ok.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>