[dv] Support extra test type for test_in_rom
Currently, tests that meant to be executed in ROM get an extra
"_rom_prog" suffix when compared to the normal ROM.

This causes a memory load error in chip_sw_base_vseq since the name
no longer matches.

This commit introduces a flag that would be recognized by
the load routine to ensure the correct version is loaded.

Signed-off-by: Timothy Chen <timothytim@google.com>