[bazel] Build disassembly with functests by default
Create per-device filegroups for all related targets in
opentitan_binary. Then reference the filegroup in the related sh_test
for the given device.

This will bring all the side outputs of opentitan_flash_binary into the
cut of the graph reachable from the sh_test (but still split on a
per-device basis).

Signed-off-by: Alexander Williams <awill@google.com>