[dv,chip] Fix a few chip level failures
Remove the pwrmgr_rstmgr_sva bound to the rstmgr instance: we already
bind it to the top_earlgrey instance, and the sva bound to rstmgr is only
suitable for unit level tests.
The fork in ast_ext_clk_if's detect_io_active_window task needs to be
removed. It clearly was a programming mistake.

Signed-off-by: Guillermo Maturana <maturana@google.com>