[dif/edn] Allow `seed_material` to be NULL for `dif_edn_instantiate()`
The documentation of `dif_edn_instantiate()` specifies that the
`seed_material` argument can be `NULL`, in which case CSRNG will use a
zero vector instead.  Prior to this commit, `dif_edn_instantiate()`
would return `kDifBadArg` in this case, however.

This commit fixes that by passing the `NULL` pointer for `seed_material`
on to `csrng_send_app_cmd()`, which in this case will correctly program
the CSRNG with a seed of length zero.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>