refactor(spi_device): Remove `bitcnt`
`bitcnt` is 12 bit data counting how many bits `spi_s2p` has received.
The signal is then sent to the following modules along with the valid
signal and a byte data.

The signal assumes the max data transfer in a SPI transaction to 256B.
256B is common for SPI Flash page size.

However, the host can read more than a page size. And the signal is
never used in the following modules (spi_readcmd, spid_upload)

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>