[dif/edn] Remove `memcpy()` in `dif_edn_instantiate()`
Prior to this commit, the `dif_edn_instantiate()` function copied the
content of the `seed_material` pointer to the stack before passing a
pointer to the copy to `csrng_send_app_cmd()`.  `csrnd_send_app_cmd()`
writes the seed array word-by-word to a register using
`mmio_region_write32()`.

It is not clear why this `memcpy()` would be required.  One reason might
be to guarantee alignment:  The `mmio_region_write32()` function "writes
an aligned uint32_t [to a memory-mapped I/O region]".  The `uint32_t`
value is passed by value, so it is present in a processor register (and
thus aligned) when the function is called.  Base address and offset
passed to `mmio_region_write32()` do not depend on the address of the
seed array, so alignment is not an issue there either.  The only source
of misalignment left is the loading of words from the seed array into a
processor register prior to calling `mmio_region_write32()`, with the
following reasoning:  C alignment requirements for `struct`s ensure that
each member variable is self-aligned [1].  The `seed_material` pointer
could be manipulated to violate this alignment, in which case the words
in the seed array would also be unaligned, and loading them could cause
address-misaligned exceptions.

To prevent this, the `csrng_send_app_cmd()` function now checks that the
`seed_material` array is word-aligned, and a unit test verifies that
`dif_edn_instantiate()` indeed rejects misaligned `seed_material`
pointers with `kDifBadArg`.

[1]: http://www.catb.org/esr/structure-packing/#_structure_alignment_and_padding

Signed-off-by: Andreas Kurth <adk@lowrisc.org>