[rv_dm dv] Fix sbcs.sberror field prediction
Based on the recent update, the sticky 3-bit sberror status
is cleared when it is written to 1. This deviates from the
standard W1C access policy behavior which clears the field
when writing the bits that are set. For example, traditionally
if status is 0x2, then a write of 0x2 would clear it. This is
not how sberror field is designed - to clear it, you would need
to write 1 instead. This commit fixes that in the RAL model
itself by creating a dedicated class abstraction for it.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>