[bazel] rename `test_binary` parameter
This updates the `opentitan_functest` macro to rename the `test_binary`
parameter to `ot_flash_binary` to better describe that this parameter
only applies to flash binaries, i.e. filegroups that are output from
`opentitan_rom_binary` macros cannot be used, and will not work with the
`test_in_rom` configuration.

Signed-off-by: Timothy Trippel <ttrippel@google.com>