[dvsim] Change global opts to only extend run opts if not duplicate
Prior to this commit, global options unconditionally extended the run
options for a test.  This could cause problems when run options already
contained an option already present in the global options, or when
global options contained the same option twice.  In such cases, the
duplicate options would be passed to the executing tool, which could
throw warnings or errors due to that.

This commit lets global options only extend run options if an option is
not already present in the run options.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>