[dvsim] Include error message cotext
The result bucketizer replaces numeric values in error
messages when jobs fail or get killed. If the job was
killed due to timeout exceeded, the error message
has the timeout value. If this error message is not
replicated in the `context` field of the ErrorMessage
class, then it is not immediately clear what the timeout
setting was.

This commit replicates error messages in various scenarios
into `context` to ensure valuable information is not lost
due to bucketization.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>