[bazel] generate unsigned and signed flash images simultaneously
There is a parameter on the `opentitan_flash_binary` macro that enables
the generation of signed images. When a `opentitan_functest` macro
passes this parameter through to the `opentitan_flash_binary`, and this
parameter was set to `true`, then only signed images would be generated.
This was an issue for chip-level tests like the UART smoke test that run
with both test ROM and mask ROM in DV sim nightly regressions as the
unsigned image would never get generated, and the testbench would look
for the unsigned image. This fixes this by generating both signed and
unsigned images when the `signed` flag is true, and only unsigned images
when the signed flag is false. This enables the same bazel label to be
passed to both dvsim.py invocations to run two different tests.

Signed-off-by: Timothy Trippel <ttrippel@google.com>