[entropy_src/rtl] Fewer alert cnt clear pulses in FW_OV mode
In order to increase predictability this commit modifies the main SM to
not clear the alert counters when finishing a SHA3 digest in FW_OV mode.

Though FW_OV mode does not send alerts for HT failures, the HTs continue
to gather statistics, and the outcome of these statistics are still
scoreboarded.  However FW_OV mode does break many of the assumptions
about this mode such as relative timing of the health tests, vs.
the SHA3 controls.   This makes it very hard to predict the
timing of the CLR pulse that gets sent to the HT fail counters.

Suppressing these CLR pulses allows the alert counters to accumulate
regardless of when FW closes off a SHA3 digest, allowing the
test environment to correctly predict the HT diagnostic counters.

Also makes a minor clarification to the DV logs to help diagnose
this problem.

Signed-off-by: Martin Lueker-Boden <martin.lueker-boden@wdc.com>