[mask_rom, status] Make `rom_error_t` compatible with `status_t`
The original mask_rom error scheme was based on specializing codes from
`absl_status_t`.  This change shuffles the bit-arrangement of
`rom_error_t` so that it is bit-compatible with `status_t`.

This change will allow further refactors, such as making the mask_rom's
`EXECUTE_TEST` macro useful across the codebase and adding LOG_ or
printf statements in tests that can display a `rom_error_t` in
textual form.

Signed-off-by: Chris Frantz <cfrantz@google.com>