[otbn,dv] Verify `START_STOP_CTRL.STATE.CONSISTENCY` CM
This commit adds a test sequence that injects faults into the handshake
between `otbn_controller` and `otbn_start_stop_control` used to request
and acknowledge a secure wipe.  When a fault is injected, OTBN is
expected to raise a fatal alert and lock up, and this expectation is
communicated to the model.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>