[entropy_src/rtl] Aggregate RepCnt & RepCnts events
This commit changes the timing of the RepCnt & RepCnts failure pulses
to make it easier to predict alerts and total failure statistics.

The challenges of interpretting and predicting the statistics of these
tests prior to this commit are described in issue #14078 (where this fix
has been proposed).

When combined with the corresponding changes to the scoreboard, this
brings the pass rate to 100% for the entropy_src_rng test.

Signed-off-by: Martin Lueker-Boden <martin.lueker-boden@wdc.com>