[entropy_src/rtl] Shift repcnt HT sampling pulse
In the repcnt_ht and repcnts_ht modules, this commit adjusts the timing
of the test_fail_pulse_o pulses to put them as close as possible to the
failing samples.  Previous to this commit the failure pulses would
be lined up with the sample after the failing sample.  This extra
delay could allow a window to pass with a failing REPCNT test.

It also creates several scoreboarding challenges in predicting
failing REPCNT outputs and statistics.

Signed-off-by: Martin Lueker-Boden <martin.lueker-boden@wdc.com>