[flash_ctrl] Qualify update with forward hints
Addresses issue found in #14056

Previously, when a buffer update is forwarded, it does not
prevent the update from happening again (with the same data)
in the subsequent cycle. Most of the time this is not an issue,
unless an allocation in the subsequent cycle happens to hit the
same buffer.

This commit adds the `hint_forward` qualification to the update
selection to ensure that if an update has already been forwarded,
there is no need to do it again.

Assertions are also added to check that updates can only happen
to a buffer that is "work in progress".  If a buffer is already
"valid", then it cannot see another update command until it has
been allocated for another transaction.

Signed-off-by: Timothy Chen <timothytim@google.com>