[otbn,dv] Fix unknown FIFO output in `otbn_sec_cm` test
The generic countermeasure test for `prim_count` manipulates the
redundant counter register to check if a fault injected into one of the
counter registers gets detected.  When this test exercises the counters
inside the response FIFO of the TLUL adapter at the instruction memory,
the `DataKnown_A` assertion in the response FIFO fails.  This is
correct, but it prevents the test from checking the countermeasure.
This commit disables that assertion for the described test case.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>