[dv/chip] Fix alert ping timeout error
This PR fixes chip_sw_alert_handler_ping_timeout error.
When alert's ping timeout is set to a small value, it might trigger
spurious alert depends on the timing. And eventually might fire an
unexpected interrupt.
To avoid this issue, this PR slightly modify the C test to check timeout
interrupt but also allows other interrupt to fire in the meantime.

Signed-off-by: Cindy Chen <chencindy@opentitan.org>