[ci] fix bug to run some smoketests in Verilator in CI
SW smoketests used to be run in Verilator in CI. However, a bug in the
shell script that runs said tests was not running them. However, fixing
this bug by running all the smoketests again increases CI runtime to
almost 2 hours.

Instead of running all SW smoketests in CI, we now run just the UART
AES smoketest with Verilator in CI as a compromise, until we can decide
what we want to run long term.

Signed-off-by: Timothy Trippel <ttrippel@google.com>