[chip dv xcelium] Fix bug in fault injection code
In the clkmgr escalation reset test vseq, we use the prim_cnt_if
sec cm proxy class to inject a fault. This is documented in our
sec cm verification infra doc page.

Xcelium does not like hierarchical referencing of class instances
declared inside an interface, and interestingly, the LRM is not
clear about this being allowed. Consequenly, VCS allows it but
Xcelium does not.

THis commit uses an alternative approach to extract the proxy class
instance - search the sec_cm_if_proxy_q defined in the sec_cm_pkg.
To achieve this, the following updates are made:
- Move find_sec_cm_if_proxy() method defined in an LC ctrl block level
  test vseq - we an reuse this at the full chip. Moved it to
  sec_cm_pkg instead.
- Made associated changes in LC ctrl block test vseq
- Invoked this method in clkmgr escalation reset test vseq at the chip
  level to find an return the proxy class instance to inject the fault.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>