[entropy_src/rtl] Suppress unneeded SHA3 start pulses
Upon disable, the `entropy_src` state machine frequently returns to the
idle state with out closing the SHA3 engine.  This is fine: it can only
increases the entropy of the following word, and is manageable in
verification.

This does however leave the SHA3 in an active state, which should
technically be accounted for.  Previous to this commit, the state
machine would generate redundant SHA3 start pulses once enabled, as the
SHA3 engine would normally be empty and idle as the SM ramps up.

Even though SHA3 engine is currently quite permissive about these
redundant start pulses, this commit provides a register to track
whether the SHA engine has already been activated to avoid
redundant starts, to prevent errors if this loophole is closed
later.

Signed-off-by: Martin Lueker-Boden <martin.lueker-boden@wdc.com>