[otbn,rtl] Relax assertions on DMEM and IMEM blanking
DMEM and IMEM are blanked during execution, during a secure wipe that
ends up in a locked state, and when OTBN is locked.  There are other
non-Idle states, though, so the `NonIdle{D,I}memReadsZero_A` assertions
are not correct.  In particular, DMEM and IMEM are not blanked during
the secure wipe after reset or during the secure wipe after execution
that does not end up in a locked state.  As argued in #14066, this is
not a problem because DMEM and IMEM are either re-scrambled in this case
(after reset) or would be readable anyway as soon as the secure wipe has
finished.

This commit relaxes the assertion so that it matches the behavior of the
RTL.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>