[entropy_src/rtl] Fine-tune SHA3 in FW_OV mode.
The PR makes a couple of minor adjustments after #14390, notably to
revert the FIFO clear mechanism in FW_OV_INSERT_ENTROPY mode.
Both changes treat minor corner cases, yet ones that come up often
in regression.

FW_OV mode requires special treatment as the FW_OV write FIFO does
not apply any backpressure when the IP is disabled (which the
RNG mode does have), thus we cannot assume that the user will
not _try_ to insert entropy during this time.  This is a narrow
corner case, and one which generates an exception, but it is
still important for the DUT to behave in a predictable fashion.

This commit also corrects the cases where the user inserts entropy
at a valid time, but then disables the DUT before the SHA3 block can
accept it.  This is technically legal, and should result in the word
being accepted and remembered by the SHA3 block.

Signed-off-by: Martin Lueker-Boden <martin.lueker-boden@wdc.com>