[entropy_src/rtl] Fine-tune SHA3 intake pipline
Since the SHA3 conditioner is not cleared on disable, it is important
to in verification carefully predict which words actually get inserted
into it as disable events occur.

This type of verificaiton problem has already been solved for the health
test statistics.   Therefore this commit lets all samples that have
already been health checked into the SHA engine, so that verification
can use the complete health-checked dataset when predicting conditioning
outputs.

This will also preserve a handful of observe-fifo words that were
previously being marked as "Dropped" by verification.  Also some bypass
words will also be preserved in the pipeline for one additional cycle as
they linger in the precon FIFO, but they will be cleared immediately
afterwards, and thus should not impact verification.

Signed-off-by: Martin Lueker-Boden <martin.lueker-boden@wdc.com>