[otbn,dv] Add missing wait for valid instruction in `otbn_ctrl_redun_vseq`
When `otbn_ctrl_redun_vseq` injects a fault into the bignum MAC op and
the instruction is not valid, no error will be raised, which is correct.
So this sequence must either wait for a valid instruction before
injecting a fault, or it must not signal an error to the model if it
injects an error during an invalid instruction.  The latter would give
more coverage, since it would also test that injected faults for invalid
instructions do not result in an error.  However, this sequence
currently waits for a valid instruction before injecting any of the
other faults.

Thus, this commit adds the missing wait for a valid instruction when
injecting faults into the bignum MAC op.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>