[chip dv] Allow SW symbol not found in elf
This commit updates the common uitl method which
searches for the SW symbol in the elf file to add
`does_not_exist_ok` argument. It is set by default
to 0.

If set to 1 and if the symbol is not found in the SW
image (elf), then it simply returns. Otherwise, if
the symbol is expected to exist in the elf and is not
found, a fatal error is thrown.

The advantage of doing this is we can unconditionally
override symbols in common {lib, testutil etc} code
that may or may not be a dependency of the test, rather
than have to know which test will be dependent on the
said lib code and condition the symbol override based
on that.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>