[sw] Enhance rand_testutils to set seed value
This commit uses the on-device entropy source to
fetch a random value to set the initial seed value
from the on-device entropy source.
for the LFSR implemented in SW. The SW LFSR
(PRNG) offers better runtime performance than
always seeking random data from the on-device
entropy source.

This enables SW randomization across all simulation
platforms. The original goal for DV was to use
a static volatile const symbol to set the seed
value, that can be overridden from the testbench.
However, after experimentation, this was found to
offer only about 3-7% runtime improvement. So,
we will use the first method for all sim platforms.

The rand_testutils provides a `rand_testutils_rng_t`
type, which provides knobs and maintains an PRNG
LFSR instance. It provides an LFSR reseed frequency
which instructs the testutil to refresh the LFSR
with a new seed after every N PRNG operations.
In addition, it provides the ability to set the
entropy fetch timeout in microseconds, and the PRNG
LFSR polynomial coefficient, which are set to
defaults upon initialization. They can be modified
if the user wishes to.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>