[ottool] Enums that can round-trip unkown values
This module adds C-like enums which can preserve un-enumerated values.
In rust code, these enums should behave as though in addition to
all enumerated values, there is also a NewType variant which is the
catch-all for unenumerated values.

The enums also have a custom Serialize/Deserialize implementations to
facilitate an optimal user experience in text serialization formats.
That is to say: for enumerated values, these enums serialize as the
named discriminant.  For un-enumerated values, these enums serialize
as the integer value.

See the doc-comments for more info.

Signed-off-by: Chris Frantz <cfrantz@google.com>