[otbn,sw] Fix data memory section for `barrett384`
The `.bss` section does not get loaded, but the `.data` section does.
The variables in this program have to be loaded though, because the test
afterwards writes only the actual data values, which is not always a
full 256-bit OTBN wide word.  If OTBN then accesses that data with wide
word operations, integrity errors are raised (also see #14552).  This is
fixed by this commit.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>