[otbn,rtl] Fix locking on RMA request while refreshing URND
Prior to this commit, when `otbn_start_stop_controller` got an RMA
request during a pre-execution URND refresh, it would start executing as
soon as the URND refresh is completed.  This is wrong: when OTBN gets an
RMA request during a pre-execution URND refresh, it must securely wipe
as soon as the URND refresh is completed (and then lock).

This bug was caused by the `if` condition on line 169 of
`otbn_start_stop_controller.sv`, which, if `rma_request` is true, is
false and thus the `else` branch on line 180 is taken.

This commit removes this bug by explicitly splitting the `if` statement
for proceeding after an URND refresh on the `stop` signal: if that
signal is high, OTBN either locks immediately or waits for the URND
refresh to complete and then securely wipes (and then locks).  The
behavior of the `else` branch is preserved.

The code is not fully DRY as the inner `else` bodies are redundant.
However, I explicitly prefer this implementation to a fully DRY one,
because it keeps the `stop` and the non-`stop` cases clearly separated.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>