fix(kmac): Not advance the ErrChk FSM in error cases
_Related Issue: https://github.com/lowRISC/opentitan/issues/14470 _

The `kmac_errchk` module reports errors in case of SW's mis-config/
mis-control. When the module detects any errors, `kmac_errchk` blocks
the SW command and report via **ERR_CODE**.

However, the module has advanced to the next state internally. The
internal FSM inside the `kmac_errchk` is to track the external and check
the correctness.

In this commit, the `kmac_errchk` stops moving to the next state in case
of `block_swcmd` occurs.

It is best to add the block condition in the `always_ff` of the `st`.
However, the sparse FSM coding style prevents it from adding the latch
enable condition. The issue
https://github.com/lowRISC/opentitan/issues/14631 has been created to
enhance the macro and the flop module.

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>