[test, sw/silicon_creator] Emit .rodata even if it's empty
extract_sw_logs_db requires .rodata to exist. Since we build for sim_dv
unconditionally and pass .rodata as an argument to that script, build
fails if a flash binary does not have the .rodata section. This commit
adds a workaround for this issue so that .rodata is emitted even if it's
empty. See #14636.

Signed-off-by: Alphan Ulusoy <alphan@google.com>