[otbn] Relax assertion for blanking of Adder Y
For `BN.SUBM` with `a >= b` (thus the result of Adder X has the carry
bit set), the result of Adder Y is not used, but it cannot be blanked
solely based on the carry bit (see issue #14603).

This commit relaxes the assertion that checks the blanking of Adder Y to
allow this case, and thereby fixes #14603.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>