[otbn,dv] Fix `OTBNState.stop()` before initial secure wipe is done
Prior to this commit, calling `OTBNState.stop()` before the initial
secure wipe has started (the model remains in `FsmState.IDLE` while it
waits for the initial URND acknowledge) caused the model to become
locked immediately.  This is different from the RTL, which in this case
correctly completes the secure wipe before becoming locked.

This can be reproduced with test `otbn_escalate` and seed `3491964481`.

This commit fixes `OTBNState.stop()` to match the correct RTL behavior.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>