[prim cdc] Disable rand delay on diff decode
This commit adds a module parameter to disable the insertion of
random delays in the CDC module. This is required for one
instance - the differentially encoded signals, which is provided
by the `prim_diff_decode` module. This module has a 1 cycle
tolerance for the differential signal being out of sync, which is
a problem for inserting the rand delay.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>