[dif] Fix dif_keymgr
keymgr status is W1C and coded like this
idle: 0
WIP: 1
Success: 2
Error: 3

Before this change,  we always read status and write the readback value.
So, when it reads WIP, then write 1 to it. Between this read and write, HW may updates
status to Error (0x3). SW clears the 1st bit and it becomes Success.

Now changed to only clear status when it's either success or error.

Signed-off-by: Weicai Yang <weicai@google.com>