Fix various typos in Markdown files
I started a pass through the repo's markdown files with `aspell`, but
stopped at `hw/lint/doc/index.md`.

I used the command below, but note that it has a few deficiencies. For
one, `aspell` does not show the name of the file you're viewing. As a
result, I sometimes found myself in vendored code or changelogs, which
should not be edited. Secondly, it's difficult to quit halfway through
as it requires exiting from `aspell` N times.

    find . -name '*.md' -exec aspell check {} \;

After the initial upload, CI complained about trailing whitespace. I
ran the suggested command on the offending files.

    util/fix_trailing_whitespace.py \
      doc/ug/directory_structure.md util/i2csvg/README.md

Signed-off-by: Dan McArdle <dmcardle@google.com>