feat(kmac)!: Add `valid` RW1C field in `ERR_CODE`
This commit adds `valid` field to `ERR_CODE` CSR. It is RW1C for SW to
clear when it acks.

It is useful for SW to check if ERR_CODE is current event or previous.

This change breaks the API as the error code field is shrinked to 7bit
from 8bit.

Related Issue: https://github.com/lowRISC/opentitan/issues/14713

Signed-off-by: Eunchan Kim <eunchan@opentitan.org>