[chip dv] Get most JTAG tests working
I totally forgot that the TAP straps are only sampled once at the start,
which is what we want for these tests. Made fixes reflecting the same.

Also added a nicer method in chip_if to set the tap straps for powerup
sampling which is not used at the moment. Test sequences will be refactored
later to use it.

This commit fixes the following tests:
- chip_sw_lc_walkthrough_dev
- chip_sw_lc_walkthrough_prod
- chip_sw_lc_walkthrough_prodend
- chip_sw_lc_walkthrough_rma
- chip_sw_lc_walkthrough_testunlocks
- chip_jtag_csr_rw
- chip_jtag_mem_access

The following tests still fail and need debug:
- chip_tap_straps_dev
- chip_tap_straps_prod
- chip_tap_straps_rma

There are more tests that use JTAG, but they are failing due to other
reasons (may be fixed in subsequent commits).

Signed-off-by: Srikrishna Iyer <sriyer@google.com>