[chip, dv] Fix chip_csr_mem_rw_with_rand_reset test
This test internally calls apply_resets_concurrently(), which was
supposed to have wiggled the POR_N. The chip level is not exactly
a CIP, yet it extends from the CIP framework, which means it
does not wiggle the POR_N (chip level uses a dedicated pins_if
for wiggling it). This commit overrides the apply_resets_concurrently()
to address that (and thus fixes this test).

Test passing with this commit:
- chip_csr_mem_rw_with_rand_reset

There are some rare seeds exhibiting unrelated failures which may
need to be triaged later.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>