[i2c] Remove scl glitch on transaction start
- addressess #14961

As part of investigation for this issue, #15003 was discovered and also
patched.

This fix just makes the clock drive low dependent on whether there is a
start bit. If a start bit is present, do not drive the clock low until
we are actually preparing a start bit. If we are just cycling through
from another path, and allow the clock to be driven low.

Signed-off-by: Timothy Chen <timothytim@google.com>