[chip dv] Fix JTAG interface enablement
THe previous code relied on the test bench setting the TAP
strap pins to enable the JTAG interface, rather than adding
a dedicated `enable_jtag` mux signal.

The existing code looked at the `tap_straps_if.pins`, which
is just a wire which will reflect the value of the pins
IOC5,8 which may have been wiggled for a different test with
a different interface, and that would inadvertantly end up
enabling JTAG, causing contention.

This commit makes it look at both, the `pins_o` and `pins_oe`
values, so that the actual intent of setting the TAP straps
is what causes the JTAG to be enabled.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>