[doc] Update motivation and objectives of DIFs
The current DIF reference manual (RM) text emphasizes error checking and
robustness of DIFs quite heavily ("making it difficult, or impossible,
to misuse [hardware]." and "providing example code to aid the
implementation of reference firmware including end-consumer
applications").  However, DIFs are called in test code, some of which
deliberately misuses a hardware module, so error checking and robustness
are not a primary objective for DIFs and could hinder their use for
tests.

This PR changes the wording of the DIF RM to de-emphasize error checking
and robustness.  The revised text is aligned with the understanding of
the Software WG at their meeting on 2022-09-27.  It also matches the
description in `sw/device/lib/dif/_index.md` more closely.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>