[sim_verilator] Correct size of flash bank
The "flash" memory area targeted bank 0 of flash, but it attempted to
claim that bank was as large as the entire flash. Correct the size, so
calls to query match its real value.

In addition, the current structure implies that the second bank cannot
be accessed via backdoor loading. Some smarter algorithm may be needed
to handle larger binaries. That is left for a future effort.

Signed-off-by: Alexander Williams <awill@google.com>