fix(spid): To not push cmdaddr for invalidlocality
_Related Issue: https://github.com/lowRISC/opentitan/issues/15251_

As described in the issue above, the TPM module uploaded a CMDADDR entry
even for an invalid locality command. The read command pointing to
invalid locality space, TPM HW returns itself with `0xFF`. There's no
need for SW to consume the command and process.

If SW wants to receive the command in any case, SW may turn off invalid
locality feature and process the command manually.

Signed-off-by: Eli Kim <eli@opentitan.org>