[top/dv] Add exit_test_unlock test.
The primary difference between this test and the existing unlock
test is that this sequence tries to simulate the liklihood system
usage.

Specifically, the ROM_EXEC_EN and the advance to PROD are done in
the same power cycle.  Upon reboot, an image is then bootstrap
into the device to ensure the behavior is correct.

The bootstrap portion for this test is still missing.

Signed-off-by: Timothy Chen <timothytim@google.com>