[otp-img] Add prod image and align partition contents with rma/dev
This adds an additional image to preload OTP in PROD state.
It also aligns the configuration contents between RMA/DEV/PROD.

Note that we may want to reduce the duplication here in the future.
This can be achieved by splitting out the partitions into separate
hjson files and feeding them to the image generation script using
the --add-cfg switch, see:

https://github.com/lowRISC/opentitan/blob/master/util/design/README.md#otp-preload-image-generator

Signed-off-by: Michael Schaffner <msf@google.com>