[bazel] update rules_rust to use current tip of tree
This updates the rules_rust Bazel depedency to use the upstream rules.
These now work in an airgapped environment.

Additionally, this has the unfortunate side-effect of temporarily
disabling rust-analyzer. To use rust-analyzer requires setting the
`include_rustc_srcs` parameter to `true` in the
`rust_toolchain_repository` repository rule invocation. However, setting
this to true, while using the nighly rust toolchains breaks airgapped
bazel builds because upstream `rules_rust` does not maintain a list of
known SHA256s for the nightly srcs. Therefore the download of the src
code that is triggered by setting `include_rustc_srcs` to `true` is not
cached.

This issue will need to be addressed in upstream `rules_rust`, and will
be handled externally.

Signed-off-by: Timothy Trippel <ttrippel@google.com>