[prim_sync_reqack] Disable reset checks by default, enable inside OTBN
This commit adds a new SV parameter to the prim_edn_req and
prim_sync_reqack(data) modules to only enable the reset checks if
explicitly enabled. For now, the checks are only enabled within OTBN.

The reason for doing this is that currently we expect quite a few test
failures based on this change. Once the reset domains have been
re-organized (see lowRISC/OpenTitan#15317) this commit can be reverted
and the checks can be enabled system wide without causing a large amount
of test failures.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>