[entropy_src/rtl] Drop corrupted SHA seeds
Fixes #15257

Specifically this follows the "Focused RTL solution" in that issue.

This commit creates a 2-bit shift register that gets set when a SHA
corruption event is detected.  This shift register is used to drop
the following two SHA outputs after the corruption event.
Once those two outputs have been dropped the shift register is clear
and further outputs are accepted.

Signed-off-by: Martin Lueker-Boden <martin.lueker-boden@wdc.com>