[util] Fix compile_commands.json for git worktrees
I tried running util/generate_compilation_db.py in a git worktree and
discovered that the "bazel-opentitan" symlink did not exist! It turns
out that "bazel-opentitan" was just a special case and the general rule
is "bazel-${workspace_name}" [0].

This PR removes the hardcoded "bazel-opentitan" and instead builds the
path relative to the source file. It moves some related path
construction into a class and adds tests.

[0]: https://bazel.build/remote/output-directories

Signed-off-by: Dan McArdle <dmcardle@google.com>