[entropy_src/dv] Refactor scoreboarding collect_entropy
When scoreboarding results from the ExtHT agent the scoreboard needs to
wait for the ExtHT agent to report back.  Since the agent is slightly
delayed by the DUT, this means that the health test scoring must also
be delayed.

This commit breaks entropy collection and health test scoring into
separate threads so that the health test scoring can be done at the
appropriate time.

Signed-off-by: Martin Lueker-Boden <martin.lueker-boden@wdc.com>