[util] Speed up generate_compilation_db.py by avoiding os.path.join
I ran the script with cProfile and saw that we were spending a lot of
'tottime' in os.path.join(), which was largely responsible for the
'cumtime' spent in reconstruct_path().

Switching to a manual join on `os.sep` saves ~8 seconds of runtime in
the profiler.

Before:
   ncalls  tottime  percall  cumtime  percall filename:lineno(function)
  1564383    3.891    0.000   12.866    0.000 generate_compilation_db.py:60(reconstruct_path)
After:
  1564383    3.411    0.000    4.480    0.000 generate_compilation_db.py:60(reconstruct_path)

Outside of the profiler, this shaves off roughly 3 seconds of runtime.

Before:
    % time util/generate_compilation_db.py --target //sw/... --out compile_commands.json
    ...
    util/generate_compilation_db.py --target //sw/... --out compile_commands.json  6.10s user 0.43s system 70% cpu 9.230 total
After:
    % time util/generate_compilation_db.py --target //sw/... --out compile_commands.json
    ...
    util/generate_compilation_db.py --target //sw/... --out compile_commands.json  3.20s user 0.51s system 57% cpu 6.452 total

Signed-off-by: Dan McArdle <dmcardle@google.com>