[test] Add AES masking off test and associated test utility functions
These test functions can be used to configure CSRNG and EDN to produce
the seed required by the AES masking PRNG to output an all-zero vector.
In combination with setting the CTRL_SHADOWED.FORCE_ZERO_MASKS bit for
AES, this allows to effectively switch of the AES masking which is
needed for ES.

This is related to lowRISC/OpenTitan#14240.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>