[dv] Coverage collection updates
The existing coverage model dumped port toggle
coverage on the ports of rv_core_ibex and its
immediate sub-modules. It is actually sufficient
to only collect toggle coverage of only the
rv_core_ibex module, because we dump coverage of
all other metrics for the complete tree under
rv_core_ibex (except the ibex_top instance).
Ports that can be covered (and are not yet
covered) will manifest as uncovered holes
in other coverage metrics (line, branch, cond
and fsm). Ports that cannot be covered since
they are unreachable will have to be manually
excluded, which is a tedious process.

This change limits the coverage collection
to only the ports of rv_core_ibex instance. We
employ the same strategy for the block levels -
we only collect port toggles of the DUT module
and not on any of its sub-modules.

Note that the port toggles on ibex_top (a
submodule of rv_core_ibex) will continue to be
dumped, becuase it is a pre-verified IP, and
we dont collect coverage of other metrics within).

The chip_earlgrey_asic and top_earlgrey only
enable coverage of itself - not its submodules.
This enables us to keep the coverage patterning
simple. This also means that any other modules
except for the major ones that we already know
of need to have full coverage enabled under them.

A cosmetic change to keep the paths sorted in
each category in alphabeical order to make
future review easy.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>