[crypto] Fix bug in ECDSA-P256 scalar inversion.
Fixes a recently introduced bug in ECDSA-P256 signatures; the scalar
inversion routine handles scalars in plaintext, so the shares need to be
combined in order to use it. Eventually we should use a more heavily
protected implementation (see #15507) but for now this just fixes the
bug by combining shares.

This commit also updates the p256 signature test to use both shares, so
that this type of bug will be caught faster in the future.

Signed-off-by: Jade Philipoom <jadep@google.com>