[entropy_src/rtl] Remove unused state from ack_sm
In the `entropy_src_ack_sm`, the ACK_IMMED state allows the
`entropy_src` to immediately output any seeds which may be available
when the module is first enabled.  Since the `entropy_src` flushes its
internal state when it is enabled, this condition is not expected, and
so this state is never expected, and appears as an FSM coverage gap.
(Note: interestingly enough, UNR analysis seems to claim that this state
is somehow still reachable, even though this condition is not expected.)

Given that this state is at best an unused optimization, it's here
removed to help close FSM coverage.

Signed-off-by: Martin Lueker-Boden <martin.lueker-boden@wdc.com>