[entropy_src/dv] Simplify recoverable alert coverage
The `entropy_src_alert` is intended to ensure that all recoverable alert
codes are detected.  However it was created before many of the required
features it tries to test were implemented in scoreboarding.

In order to enhance maintainability, this commit updates this test to make
better use scoreboarding and of the configuration features in the
entropy_src_base_vseq.  It also adds a new covergroup to catch all the
recoverable alert conditions not included in the `mubi_err_cg`.

Since almost of the `recov_alert` coverpoints can be accessed by the
`entropy_src_rng` test and confirmed in scoreboarding, most of the
`entropy_src_alert` test has been trimmed down to just perform the tests
that can not be performed elsewhere.

The `entropy_src_alert` testpoint has also been renamed to
`entropy_src_functional_alerts`, to highlight the difference with the
`entropy_src_alert_test` which tests the alert system through the CSR
test interface, but does not replicate the actual functional conditions.

In addition to streamlining the `entropy_src_alert_vseq` this
test simplifies the `entropy_src_smoke_vseq` to use basic
`push_pull_host_seq` (as opposed to the indefinite variety).

Signed-off-by: Martin Lueker-Boden <martin.lueker-boden@wdc.com>