[fpga] Fix OTP-specific behavior in MMI file generator
I discovered a few problems with the OTP-specific MMI files we have been
generating prior to this commit.

(1) The file's bitlanes were overlapping. For instance, the first
bitlane claimed to cover bits 0-3 (inclusive) and the second one claimed
to cover bits 1-4. This is expressly not allowed by Vivado's updatemem
manual, so I knew something was wrong. By poking around the Vivado GUI,
I realized that OTP bitlanes should each cover a single bit. I'm not
sure whether that logic is correct or necessary for the ROM case, so I
left it alone.

(2) The MMI file claimed that BRAMS are the type "RAMB32" when the OTP
BRAMs are actually RAMB18E1. It's not as simple as just saying "RAMB18"
in the MMI file. Ultimately, I got it working after bumping the MMI's
minor version from 0 to 1 and configuring an environment variable before
calling updatemem.

(3) The bitlane "Placement" attributes were all wrong because we were
trying to trim the "RAMB36_" prefix from strings that actually began
with "RAMB18_".

(4) The AddressSpace and and DataWidth tags had inaccurate values.
Apparently there is a significant amount of padding between OTP words in
the OTP BRAMs. In fact, each word is followed by 15 zeroes.

Separately, I also modified the TCL script to dump the OTP BRAMs'
INIT_XX strings to a text file. Ultimately, examining these INIT_XX
strings helped me crack the way that Vivado lays out data in the
BRAMs (more detail in a later commit). However, examining the INIT_XX
strings in the Vivado GUI required too much clicking; there are 64
INIT_XX strings per OTP BRAM, and there are 22 BRAMs. This is purely a
debugging aid, but it only takes a few seconds to generate.

Signed-off-by: Dan McArdle <dmcardle@google.com>