[crypto] Reorder AES driver checks to avoid race condition.
The AES driver documentation specifies that the `dest` pointer should
accept the output from the *previous* input block (if any), not the
current one. However, the implementation actually passed the next block
of input before reading the output, which could lead to a race condition
where the AES block overwrites the output data before it is copied to
`dest`. This has not been triggered by any tests but it seems better to
be on the safe side and fix it preemptively.

Signed-off-by: Jade Philipoom <jadep@google.com>