[dv/chip] Fix prim_tl_access xcelium failure
This PR fixes prim_tl_access xcelium failure due to the following
scenario:
In dut_init, the stub_cpu sequence tries to write ast's csr. But in
xcelium, it happened before the ast_tlul_rst is set to 1. So the write
is ignored.

The fix is to wait for the reset to set to 1 before doing the CSR write.

This fixes issue #15459.

Signed-off-by: Cindy Chen <chencindy@opentitan.org>