[otbn,rtl] Latch various internal state errors
FPV has revealed it was possible for some internal state errors to be
ignored, so no alert was triggered, if they occurred the same cycle or
the cycle before a start command was issues to OTBN. This was due to
some errors only appearing for a single cycle and others getting reset
by the start command.

This adds latching behaviour to various internal errors, so once seen
they will remain asserted until reset.

Signed-off-by: Greg Chadwick <gac@lowrisc.org>