[rom] Fix pinmux indexing and set attributes first.
The pinmux initialization previously used the pinmux input selector
value as the pad index, which results in an off-by-two error since the
input selector is (pad index + 2).

Also, the pinmux documentation specifies that pad attributes should be
set first, before configuring the pinmux matrix, so this change moves
attribute-setting to precede IO configuration.

Signed-off-by: Jade Philipoom <jadep@google.com>