[test] Add DV component for AES masking off test, add it to the testplan
This commit adds the AES masking off test to the testplan. As the
masking itself is transparent to software, a DV component is added to
probe into AES and verify that 1) all prerequisites are met to switch
off the masking and 2) that the second share of the state as well as
mask input and output of SubBytes are all zero, i.e., masking is indeed
switched off.

This is related to lowRISC/OpenTitan#14240.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>