[top-level,test] add initial power virus test boilerplate
This adds the boilerplate code for the power virus test to partially
address #14814. Since this test will be rather large/complex, it will be
committed in pieces over time. This initial commit contains a basic test
checklist (that will be added-to/improved over time) and basic
configurations for the pinmux, gpio, and adc_ctrl peripherals.

Signed-off-by: Timothy Trippel <ttrippel@google.com>