[top/usbdev] Make USB_P/N pads direct instead of manual
In the past, we had to connect the USB_P/N IOs to two physical pad
instances each for electrical reasons, which required these signals
to be `manual`. By now, we have moved these dual pad instances into
the pad wrapper, which means that there is no reason to keep these
connections `manual` anymore. They can be made `direct` so that the
connections are made automatically at the chip-level.

Signed-off-by: Michael Schaffner <msf@google.com>