[entropy_src/dv] Remove bad call to uvm_reg.reset()
The "reset()" method for uvm_reg does not actually reset or even
attempt to reset the DUT's register, only the mirrored value and
associated mirror state.  This is a problem if the register is locked
from the hardware side.

This commit removes a redundant call to reset, that apparently had
already been supplemented with a call to `csr_wr(ral.<reg>.get_reset())`

Signed-off-by: Martin Lueker-Boden <martin.lueker-boden@wdc.com>