[chip dv] Deprecate chip_common_pkg::chip_io_e
This commit removes the `chip_io_e` enum that was hand-written in
chip_common_pkg, in favor of the enums that are already generated in
top_earlgrey_pkg.

This change splits the all-encompassing `ios_if` and the accompanying
`ios` signal in `chip_if` into two separate `d|mios_if` and `d|mios`,
for dedicated and muxed IOs. This eases the pinmux / pad related
testing.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>