[entropy_src/dv] Better cross coverage of FW_OV parameters
Though the `FW_OV_ENTROPY_INSERT` setting should do nothing when
`FW_OV_MODE` is false there are many cross coverpoints that look for
this combination.   This update to the `dut_cfg` constraints randomly
picks a 50-50 value for `FW_OV_ENTROPY_INSERT` when `FW_OV_MODE` is
false, regardless of the other knobs say, simply to get coverage
closure, and confirm that this behaivior is obeyed.

Signed-off-by: Martin Lueker-Boden <martin.lueker-boden@wdc.com>