[aes, dv] Apply resets concurrently when recovering from fatal alerts
Previously, only the main reset was asserted which sometimes caused
assertions in the push/pull agent to fire in case the DUT drops the EDN
request because it's getting reset. To avoid this, both resets are
asserted simultaneously when recovering from fatal alerts as during EDN
reset, the assertions inside the push/pull agent are disabled.

This resolves lowRISC/OpenTitan#13573.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>