[test] Implement rom_e2e_asm_watchdog bite tests
Test that the watchdog's bite resets the core. This is achieved outside
of GDB, by inspecting its output.

This commit adds the --gdb-expect-output-sequence flag to the GDB test
coordinator script. When the script sees GDB print each of the given
lines, in sequence, it will terminate GDB on the spot and consider it a
success.

This enables us to assert that GDB reaches an echo line from the GDB
script and *then* sees the core reset.

Signed-off-by: Dan McArdle <dmcardle@google.com>