[entropy_src/rtl] Proper checking of fw_ov_wr_fifo_full
In `es_bypass_mode` FW_OV data goes to directly to the bypass FIFO, not
to the precon fifo.  This commit updates the fw_ov_wr_fifo_full register
to reflect that.

In a related vein, the es_bypass_mode signal now updates immediately when
even fips_enable, es_route or es_type are updated so that changes in
the states of this signal can be tracked in FW.

Signed-off-by: Martin Lueker-Boden <martin.lueker-boden@wdc.com>