[dv] Always preload ROM in chip_base_vseq
chip_stub_cpu_base_vseq which caused the ROM init done check in
dut_init() of the chip_rv_dm_lc_disabled_vseq to hang forever.

This PR removes the ROM init check from chip_rv_dm_lc_disabled_vseq
to fix that. In addition, it always pre-loads a valid ROM image
during the pre_start() phase of chip_base_vseq to make sure that
the ROM check can succeed successfully. Note that the real hardware
will always contain a valid ROM image, so this is not a simulation-only
hack and bring simulations closer to the behavior of the real hardware.

This fixes #15893.

Signed-off-by: Michael Schaffner <msf@google.com>