[rdc/spid] Waive CPHA, CPOL metastability
If CPHA is set, MainReset, SwRsqRst, POK reset causes a single edge to
SPI_DEV_IN or SPI_DEV_OUT clock. As the edge occurs at the same time
with the IP reset, it may create metastability issue to many registers
inside SPI_DEVICE IP.

However, at the time when POR is released, or SW requsts the reset, SPI
is assumes as quiescent or stalled. If SPI line is stalled, part of the
internal content may not be reset. But, it can be solvable by setting
invert to the CSb pad.

Signed-off-by: Eli Kim <eli@opentitan.org>