Update `Install Vivado` for the 2020.2 installer.
The 2020.2 installer moved the EULA agreement step later in the install process. This PR reorders the Vivado installation guide's instructions to match the 2020.2 installer, so it is easier to follow. I renamed all the images to use descriptive names, so future updates are less disruptive (e.g. if Vivado decides to reorder steps again).

I provided a new version of the features-to-install screenshot, as 2020.2 has some options that 2020.1 did not.

Last, I added a note that you will want to add Vivado's `bin` directory to your `PATH` variable so that Bazel can invoke `updatemem`.

Signed-off-by: Johnathan Van Why <jrvanwhy@google.com>