[dv] enable tests to specify a custom OTP image built by Bazel
Previously, only a specific set of five different OTP images (built by
dvsim during the pre-run phase) were avaiable for tests to use. If
a test required specific OTP fields to be different than what was
provided in any of these base image, the test needed to backdoor
overwrite said field.

To enable running ROM E2E tests in DV, which have their own Bazel
infrastructure to generate OTP images as part of the test build process,
this enhances the DV testbench to enable passing a custom OTP image
(which is automatically built by Bazel when the SW images are built for
a test during the dvsim run phase). This allows us ROM E2E tests to use
the same OTP image across all platforms (FPGA, Verilator, and DV), with
the exception of the img- and otp-seed values which may be specified on
the command line when Bazel is invoked to build the OTP image (thanks to
the previous commit).

Signed-off-by: Timothy Trippel <ttrippel@google.com>