[entropy_src/rtl] Consistent core enable masking
The entropy_src core module has many activities that persist after
disablement, notably the closing of health_test statistics data
and the processing of SHA conditioning outputs.  Prior to this
commit there have been many ad-hoc signals to allow these individual
processes to finish before clearing the relavant fifos or state
machines.  This lack of consistency makes it hard to track
how these delays interact.

Ideally the external module interfaces (CSR locks and RNG inputs)
would also be closed during these shutdown period to prevent
the admission of new RNG data (bad for verification) or new CSR
commands (which can cause spurious events due to poorly timed
changes in configuration).

This commit creates a new delayed copy of the module enable
signals.  The falling edge of this delayed copy signals when
it is safe to clear the last few internal FIFOs, and the rising
edge signals when the FSM has actually received a pending idle
request.

This new enable signal is generated by a new module in order
to cleanly encapulate and document the logic and motivation for
the delay calculations.

Signed-off-by: Martin Lueker-Boden <martin.lueker-boden@wdc.com>