[dv/edn] support csrng agent disablement
When the EDN is disabled, we expect the CSRNG to also disable.
The current CSRNG agent does not know that the CSRNG is disabled, and
will still try to response to the current CSRNG request (dropped due to
disablement).
So I connect the csrng agent's reset. This will help drop the CSRNG
request.

Signed-off-by: Cindy Chen <chencindy@opentitan.org>