[test] Extend the chip_sw_data_integrity test
This extends the chip_sw_data_integrity test so that it also
injects errors in the retention SRAM and into instructions of
a test program in the main SRAM.

The main test mechanism remains the same, but the ISRs and exception
handlers are restructured in order to perform more strict checks.

Addresses part of #15970 and part of https://github.com/lowRISC/ibex/issues/1910.

Signed-off-by: Michael Schaffner <msf@google.com>