[csrng,dv] Improve scb and sample new CGs in tests
Since we generally use backdoor accesses to CSRs, only using tlul
access related scoreboard sampling will not be enough. Hence, we
sample the covergroups with their expected values after the check.

That way, if expected value won't match we won't get the coverage
but in the case of a match we don't need to do clever tricks to
sample in an appropriate time as well.

Signed-off-by: Canberk Topal <ctopal@lowrisc.org>