[dv/pwrmgr] Fix bug in objection count handling
The slow clock io_clk_en pipelines stopping or starting the clock
and updating io_clk_val. These actions need to be in separate threads
or some changes weill be missed if they happen in close temporal
proximity.
Instrument the objection count updates so it is easier to detect
problems.

Signed-off-by: Guillermo Maturana <maturana@google.com>