[csrng, dv] Extend csrng_alert_vseq.sv to also test CTRL.MUBI
This CM was previously untested. The test now also verifies that
upon writing an Application Interface Command Header for an Instantiate
or Reseed command with invalid MUBI value in the FLAG0 field on one of
the application interfaces, a recoverable alert is fired and the
corresponding bit in the RECOV_ALERT_STS register is set.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>