[dv/pwrmgr] Fix recent unit test failures
Some of the fixes:

The change in clock for alert handling being clk_lc rather than clk (and
the same change for reset) caused the cip scoreboard to loose track of
alert expectations. Rather than changing cip, this creates a couple of
functions that enable the pwrmgr to assist cip track resets, and adds
calls to these in various test sequences.

The aborted_low_power test was unreliable, so this changes the test to
rely on internal state machines to determine when to send stimulus and
to predict expected results.

Signed-off-by: Guillermo Maturana <maturana@google.com>