[i2c] Update stretch handling
Change the way stretch is handled from "before" an event to "after" an event.
Prior to this change, the stretch check was done before start / address / byte processing.

This created a few unnecessary scenarios around start / address and made it more difficult to
add optional nack support later.  By moving the stretching decision to after
an event, we can more easily align it with nack later on since the fork can be done prior to
sending the ack bit.

AFTER in this case means we always receive the address or data byte. Upon receiving the byte,
we then inspect to see if there is space in the FIFO.  If there is not, we stretch the ack
bit until space is available, then we desposit the format byte and move on.

When nack support is added in the future, we can then simply make a decision to send back a
nack bit instead of stretching, otherwise the logic flow should be nearly identical.

Signed-off-by: Timothy Chen <timothytim@google.com>