[i2c] Restructure acq to be able to support Hs in the future.
While Hs mode is not currently supported, the previous stretching
organization made it difficult to support Hs in the long run.
Minor tweaks to addr and acq stretching flows to better support this.

The stretching scheme works as follows:
For data input (address / write operations), the stretch is always done
after the data is ack'd.  For data output (read operations), the stretch
is done before the data is sent.

This means for address / write operations, we always receive the byte,
acknowledge the byte and then look to see if space is available in the FIFO.
If space is not available, the clock is stretched until software makes space.
Acq full interrupt is used to notify software.

For read operations, we always look to see if there is a byte to send first
(functionally after the last ack from the host).  If there is nothing to send,
the clock is stretched until software provides data.
Tx empty interrupt is used to notify software.

Signed-off-by: Timothy Chen <timothytim@google.com>