[rdc] USBDEV wakeup req
USBDEV has WAKE_CONTROL CSR to initiate the wake up detector inside
usbdev_aon_wake module (inside pinmux). The CSR is write-only. When the
main power is off, the USBDEV CSR is being reset. The wakeup module
inside the pinmux IP directly uses the signal, which creates the
metastability issue.

However, the signal is not active when the power becomes off. The error
can be waived.

On the other hand, there is a lurking bug in the module, the
metastability may be ignored, but the reset may introduce the
discrepancy between the USBDEV and the pinmux as the reset only clears
the USBDEV IP.

Signed-off-by: Eli Kim <eli@opentitan.org>