[spi_device] Synchronize FIFO depths at ends of transactions
Do not show software partial transactions in reports of FIFO occupancy
counts. Instead, maintain a separate count from the FIFO, and
synchronize with the raw count from the FIFO only at the end of a
transaction. This will prevent software from getting confused about
whether a command is ready to be processed.

Signed-off-by: Alexander Williams <awill@google.com>