[dv] continue sim even if SW logs database files are missing
Some ROM E2E tests load corrupted software images that have no SW
logging database files. This simplifies the testbench to not crash if
these files are not present. Additionally, this removes the need to add
the "no_sw_logs_db" flag to such images in the dvsim config file.

This addresses review feedback in #16335.

Signed-off-by: Timothy Trippel <ttrippel@google.com>