[rdc] Turn off SPI_HOST when reset
For SW_RST_REQ, AonPok, and MainPok scenarios, the device is assumed to
be in idle state. One common error shows up in the report is from
SPI_HOST CONTROL.output_en to SPI_HOST PADs.

This commit adds constraints to turn off Passthrough and the SPI_HOST
Output when resetting the IP.

Signed-off-by: Eli Kim <eli@opentitan.org>