[aes] Improve FI detection when OR-combining multi-rail signals
Previously, OR-combination and error detection for multi-bit, multi-rail
signals was done in a single for loop inside an always_comb block. In
every loop iteration, the current combined signal was only compared to
the current signal added to the combination. This is not ideal from a
fault detection perspective.

For example, when faulting the last rail of the round counter to 1 when
the counter is actually 0 wouldn't be detected immediately. For the
first and second loop iteration, the OR-combined value was still 0. For
the last rail, the OR-combined value would be 1 which is equal to the
faulted counter. In practice this didn't really matter because a faulted
round counter itself isn't problematic. But the faulted round counter
changes the control flow of one rail which is always detected, e.g., by
a register write enable signal taking on an invalid encoding as the
different rails driving that get out of sync.

This minimal RTL change performs the comparison in a separate loop to
effectively compare all rails with the total OR-combined value. In the
example above, the final OR-combined counter value is compared against
both un-faulted rails and the fault is immediately detected.

The associated area impact of this change is neglibile. With Yosys +
nangate45, AES is now 110.17 kGE vs. 109.86 kGE (+0.28%). But it
improves FI detection and most importantly simplifies DV as faults in
the round counter are detected immediately and independent whether
the AES cipher core is busy or idle.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>