[rdc] Add quiescent DMI req/rsp constraints
Current prim_fifo_async has one bit pointers inside if Depth is 1.

If JTAG TRST is asserted when the pointer is 1, then there's chance the
metastability issue occurs on the core clock side.

One way to resolve this is to make the async FIFO synchronous reset. It
is not as of now (async assert / sync de-assert for JTAG TRST) in
dmi_cdc.

Either we can waive the scenario as asserting TRST_N during the JTAG
debugging is rare in my opinion, or making synchronous reset in dmi_cdc
would be solutions.

Signed-off-by: Eli Kim <eli@opentitan.org>