[dv/pwrmgr_ast] Fix pwrmgr_ast SVAs
When the power glitch occurs the AST reacts by changes in *clk_val.
Previously there was an assertion requiring changes in the pwrmgr
*clk_en outputs, which were violated.
Also, the assertions are really edge triggered by changes on por_n_i.
The *clk_en to *clk_val handshake are also triggered by edges.
The maximum number of cycles to wait for transitions is increased.
The macros local to this file are undefined before the end of file.

This fixes one of the failure scenarios in #14164.

Signed-off-by: Guillermo Maturana <maturana@google.com>