[csrng/rtl] Fix disable value of `entropy_src` FFs in `csrng_core`
The `entropy_src_seed_q` and `entropy_src_fips_q` FFs are currently not
cleared on disable.  This means their value for the first command is
wrong, as the RTL relies on them being zero to correctly handle
*Instantiate* commands that do not use entropy from `entropy_src`: The
first *Instantiate* after re-enable does not have `cmd_entropy_avail`
yet, and if `flag0[1]` is set, the previous value of the seed and the
FIPS bit in those registers gets used instead of all-zeroes.  If the
previous value is also all-zeroes, there is no difference and this case
is handled correctly.  If the previous value is non-zero, this case is
handled incorrectly.

This commit fixes that problem by making the value after re-enable of
those registers equal to their value after reset.

Signed-off-by: Andreas Kurth <adk@lowrisc.org>