[csrng_agent] Correct handling of invalid commands for coverage
This commit fixes the sampling of illegal commands in `csrng_monitor`:
the DUT does not respond to these commands, so the monitor must not wait
for the ack in case of illegal commands.

This commit also refines the coverage crosses defined in
`csrng_agent_cov`; `csrng_cmd_cross` is split into two crosses:
- `csrng_cmd_clen_flag_cross` tracks the coverage of commands and their
  parameters (i.e., `clen` and `flag0`), and
- `csrng_cmd_clen_flag_sts_cross` tracks the coverage of commands,
  parameters, and the returned status.

The first cross tracks coverage of all command-parameter combinations,
and the second cross tracks coverage of the expected status responses
for all command-parameter combinations.  In the latter, illegal commands
are excluded because their status is not defined (as there is no
response to illegal commands).

Co-authored-by: Andreas Kurth <adk@lowrisc.org>
Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>