[fpv] Support expected failure properties
This PR supports parsing results with expected failures.
Some of the formal properties are expected to failure in normal formal
testbenchs.
For example: most of the security countermeasures are not supposed to be
proven under normal formal conditions. They are only reachable via
forcing the env.

So this PR adds support to results parsing. It allows user to input a
expected_failure hjson file. And these failing properties won't be calculated
towards total failing tests.

Signed-off-by: Cindy Chen <chencindy@opentitan.org>