[dv] sec_cm sequence changes to provide more control
These changes add the ability to disable specific fault injection
proxies and for a fault injection proxy to fail to inject a fault.

The FSM proxy uses the fault injection fail ability to try to inject an
FSM fault on specific states. It times out and fails to inject the fault
after a configurable number of cycles.

Signed-off-by: Greg Chadwick <gac@lowrisc.org>