[rom_e2e,dv] apply reset after boot fault in vseq
When a ROM E2E test boot faults, an alert is generated. Ending a test
when an alert is fired triggers an assertion. This updates the ROM E2E
base sequence to apply a reset before ending the test to enable better
test portability in the future. Additionally this addresses review
feedback in #16335.

Signed-off-by: Timothy Trippel <ttrippel@google.com>