[i2c] Allow tx_fifo to self clear upon transaction complete
Currently, if tx fifo data is not completely drained during a
target read, a nonempty interrupt is fired to alert software.

However, if before software handling is available another
read is started, it is possible we will send out stale data
(imagine for example the reads are targetting different registers).

In order to get around this, allow the tx fifo to self clear upon
completition, this will cause the target to stretch and give software
ample time to complete.

Signed-off-by: Timothy Chen <timothytim@google.com>