[dv/edn] Fix regression illegal bin failure
this pr fixes regression failure regarding the recent update on CSRNG
agent illegal bins.
The `sts` is a random input for EDN and thus agent can randomly drive
true or false.
There is a constraint in CSRNG side that sts will also be pass if it is
genbits command. But since it has no effect on EDN side, I think it
might be better to keep it driving random values.
So I separate coverage collection for host and device agent.

Signed-off-by: Cindy Chen <chencindy@opentitan.org>