[chip dv] Add `configure_uart_agent` function
THis function is repeatedly used in several sequences. It
makes sense to create a reusable function.

This function has sufficient args to configure the UART agent
exactly how the chip device will expect it.

Note that there are several sequences that can call this function.
Those will be updated in a subsequent PR.

Signed-off-by: Srikrishna Iyer <sriyer@google.com>