[opentitantool] Use PullMode instead of WeakPushPull
The output pins of HyperDebug or simulated "environment" can be employ
strong drive, either push/pull or open drain, and separately can enable
weak pullup / pulldown in case the strong drive is not active (e.g. when
output level is high, and the mode is "open drain", or when the pin mode
is "input".)

These options should be enough to simulate all weak and strong
strappings, provided that Toshak's claim that HyperDebug "weak" pull is
going to be weaker than the weak pull of the future OpenTitan ASIC
holds.  But for now, we can make sure our host_emulation and verilator
environments behaves such that the weak pull of the simulated
environment is weaker than the simulated OpenTitan chip.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I6aeb8d52bbe92227695a7e24c4bab93f00f8570b