[dv] Address #6628
The issue #6628 suggests disabling TLUL assertion based on the TLUL interface
that it's being tested.

This is the path of tlul_assert.
> tb.dut.tlul_assert_device.gen_device

If we want to support it, we have to enforce a rule that users use `ral_name` as part of
the tlul_assert bind name, for example
`bind uart tlul_assert#(...) tlul_assert_device_uart_reg_block(...);`

I don't think it's worth doing that, since we only disable tlul_assert when we test tl_error
and tl_intr_err. Perhaps don't make things too compicated.

Hence, I would like to remove the TODO and close the issue.

Signed-off-by: Weicai Yang <weicai@google.com>