[opentitantool] Atomically modify mode/value/pull
The current gpio trait cannot avoid glitches, if a configuration file
declares that a certain pin needs to be e.g. open drain and high logical
level.  If the new level is applied before the open drain mode, and the
pin happened to be in push/pull mode before, strongly pulling down, then
it will briefly strongly pull up, before the open drain mode is applied.
Conversely, if the mode is applied before the level, and the pin
happened to be in input mode, with the unused "output value" being low,
then the pin will briefly strongly pull down, before being released.

It is desirable to avoid such glitches.  This CL lays the foundation for
that, by adding the means to simultaneously tell the transport backend
that the mode, output level, and weak pull mode needs to be modified.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I552016571af525cdb872f98abaf915c64ed57563