[kmac] Make masked software keys optional for the unmasked design
This commit adds a new top-level parameter to KMAC to optionally enable
masking of the software key in case the unmasked design is chosen. This
is useful for optionally having fully compatible software interfaces
between the masked and unmasked design but not unnecessarily inflate the
area of the unmasked design.

In case masking is enabled, the parameter has no effect. Software then
always provides the key in two shares.

This resolves lowRISC/OpenTitan#16206.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>