[dv/otp] Fix otp_ctrl_init_failure regression error
This PR fixes the otp_Init_failure regression error where:
the sequence tries to trigger a LC transition failure by clearing the
memory, then do a manual check and expect a LC partition failure.
But if the partition is previously locked, this will trigger a HW
partition failure before the LC partition failure.

Signed-off-by: Cindy Chen <chencindy@opentitan.org>