[top/dv] Make alert checking in Ibex lockstep glitch test more robust
Instead of waiting for a fixed number of cycles and checking for a
potential alert in a single clock cycle, we now continuously check for
potential alerts during a window of max_delay_clks clock cycles after
injecting the glitch. This helps to make the test more robust.

This resolves lowRISC/OpenTitan#16954.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>