[top/dv] Fix waiting for instr/data_rvalid_i usage in Ibex lockstep test
The rvalid signals are used by the design one clock cycle after the req/gnt
handshake and the test logic for tracking outstanding accesses always
updates on the negative edge. Previously, the test erroneously applied the
glitch on the same negative edge when detecting the handshake. As a result,
some glitches weren't actually effective even though the test expected this
and thus failed.

This is related to lowRISC/OpenTitan#16758.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>