[bitstreams] Use the bazel-built ROMs by default
Change the default to use bazel-built ROMs for bitstreams instead of
whatever was in the bitstream when it was uploaded to the cache. This
makes the default bazel flow use the ROMs that are part of the current
source tree and avoids relying on a particular variant being available
in the cache.

As a side effect, software development on the FPGA now firmly requires
the FPGA lab tools by default. Update the documentation to reflect the
new behavior and tool requirements.

Signed-off-by: Alexander Williams <awill@google.com>