[crypto] Fix a race condition in the entropy driver.
I believe this fixes a bug that caused unpredictable behavior in the
entropy driver when timing was changed (e.g. tests that passed or failed
depending on whether a log statement was present). The entropy driver
docs say we must wait for an ack before issuing new commands, so the
problem was potentially caused by writing a new command before the ack
bit was present.

Signed-off-by: Jade Philipoom <jadep@google.com>