[opentitantool] Updates for final HyperDebug firmware
The OctoSPI support of HyperDebug firmware has now been merged into the Chrome OS codebase.  In
the process, a few adjustments had to be made, which unfortunately required changes to the
text-based protocol with the host machine.  This CL makes corresponding changes to
transport/hyperdebug, with fallback to handle older firmwares.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I562b64110ce13f300d6662ebea2577461433e397