[dv] Exclude prim_secdec_* in coverage collection
These prim_secdec* are verified in FPV
prim_secded_inv_64_57_dec and prim_secded_inv_39_32_dec are used in reg_top for checking integrity values.
They have many conditional coverages. It's hard to hit 100% and some of them are unreachable.
It's better to exclude them for all the blocks, since it's unlikey to miss any bug in these 2 prims.

Signed-off-by: Weicai Yang <weicai@google.com>