[sram_ctrl] Fix RMW corner case when escalating
If a global escalation request and a RMW transaction
come in at the same time, the RMW transaction will not
be able to complete even though the TL gate allows for
transactions that are already pending inside the device
to respond back (to make the bus behavior less erratic
for a better SW debug experience).

The reason for this is the write part of the RMW
transaction, since the scrambling device blocks
any transaction in case the keys have been invalidated
due to escalation.

This patch fixes this corner case by specifically
allowing pending writes to complete upon escalation.

The rationale is that a write won't give an attacker any
advantage, especially since the scrambling keys are already
clobbered at that point, and the TL gate is not accepting
any additional transactions from the bus.

Fixes #17064.

Signed-off-by: Michael Schaffner <msf@google.com>