[opentitantool] Lock serial port before using
By default Unix character devices can be opened by multiple processed
simultaneously.  For serial ports, that results in the arriving data to
be arbitratily divided between the multiple readers.  It is hard to
imagine a scenario on which that would be desirable.  When done by
accident, by e.g. forgetting about a `minicom` in the background, when
starting `opentitantool console`, the result could be that a few
characters seem to be "dropped" now and then (when really those few
characters appear in the background `minicom`), causing engineers to
waste time looking for loose cables, or firmware bugs, when none exist.

For that reason, programs such as `minicom` will attempt to atomically
create a file under `/var/lock`, to prevent multiple instances from
opening the same serial port.  (See specification of the exact file
format here:
https://refspecs.linuxfoundation.org/FHS_3.0/fhs/ch05s09.html )

Another terminal program, 'tio` (used by the author), uses the Linux
system call `flock()` on the serial device itself for the same
effect (with the unfortunate result that one `minicom` instance and one
`tio` intance could simultaneously open the same port.)

This CL makes OpenTitanTool BOTH use `/var/lock` and `flock()`, to
prevent accidentally opening the same serial port with `opentitantool
console` that is used by either other `opentitantool` instance, or by
`minicom` or `tio`.

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I62861c230835991312664b43b493e3ec5f42c314