[top/dv] Always glitch on negative clock edge in Ibex lockstep test
This commit changes the Ibex lockstep glitch test to always wait for at
least 1 negative clock edge before inserting the glitch. Previously, it
could happen that the test only waited for a specific sw_test_status.
Changes to this variable always happen on the positive clock edge.
As a result, the glitch was erroneously inserted on the positive clock
edge and removed again after half a clock cycle. Some glitches may not
be effective this way.

Signed-off-by: Pirmin Vogel <vogelpi@lowrisc.org>