[base] Correct error in declaring module ids
The `status.h` module declared `kStatusModuleId` as an global and intended
users to use `DECLARE_MODULE_ID` to create a local override with `static`
scope that would shadow the global.  Embarrassingly, I never wrote an
explicit test for this override behavior and the compiler complains
that a variable cannot be declared both `extern` and `static`.

In order to permit this shadowing behavior, I've updated the code to
name the default module ID as:
```
extern const uint32_t MODULE_ID;
```

and then permit users to `#define` their own MODULE_IDs to provide the
relevant constant (preprocessor substitutions always win):

```
 #define MODULE_ID MAKE_MODULE_ID('a', 'b', 'c')
```

Signed-off-by: Chris Frantz <cfrantz@google.com>