[opentitantool] Add provisions for updating debugger firmware
The STM32L552 chip comes with a ROM bootlader capable of programming its
internal flash with a firmware received via USB.  This CL allows
`opentitantool` to use that protocol to update the HyperDebug firmware.

Usage, if the board is running recent HyperDebug firmware:
`opentitantool --interface hyperdebug transport update-firmware`

Usage, if the board is already in DFU bootloader mode through BOOT0
strapping having red and blue LED on.  This is the way to program a
Nucleo board fresh from the factory:
`opentitantool --interface hyperdebug_dfu transport update-firmware`

Signed-off-by: Jes B. Klinke <jbk@chromium.org>
Change-Id: I095477f52e5f98357c05342e99fcdc716ac8ba37