[coverage] Use non-instrumented test ROM
Due to the recent changes in bazel targets, we now splice locally built
ROMs into the bitstreams from the bitstream cache by default. However,
when coverage is enabled, the instrumented test ROM overflows the space
allocated for ROM. This commit
- Modifies `handle_test_targets()` for functional tests so that the FPGA
is programmed with the non-instrumented test ROM before the instrumented
tests are executed, and
- Modifies `measure_coverage()` so that we don't program the FPGA while
running the tests.

Signed-off-by: Alphan Ulusoy <alphan@opentitan.org>