[dv/sysrst_ctrl] Fix nightly failure
This PR fixes sysrst_ctrl nightly failure because an index issue.
According to `key_intr_ctl` register, the first set is index 0 to 6, the
second set is index 7 to 13.
So in a for loop, if the first set is `i`, the second set should be
`i+7`, but the implementation is `i+7+1`.
This PR removed the `+1` and also add two begin end to improve the
readability.

Signed-off-by: Cindy Chen <chencindy@opentitan.org>