[otbn] Stop OTBN URND when it is not used
Stop advancing URND when data is not required by the program.
The purpose of this change is to reduce noise for side-channel analysis
of OTBN.

This feature is hidden behind a parameter SecMuteURND and it should be used
only for generating bitstreams for side-channel analysis.
An assertion is added to ensure that this feature is not enabled by accident.

Signed-off-by: Vladimir Rozic <vrozic@lowrisc.org>