[dv/kmac] Fix nightly regression error
This PR fix issue #17094

The strong liveness assertions in sha3pad module are written to ensure a kmac request/transaction
must finish. However, in chip level testbench, it is not necessary to wait for all kmac
transactions to finish. A recent update in VCS will flag these unfinished strong assertions as
failures.
So the solution here is to move these assertions to block level testbench, using an interface with
binding method. This gives user flexibility to choose whether they want to enable these strong
assertions.

Signed-off-by: Cindy Chen <chencindy@opentitan.org>