[util] Add fix_imports to enable absolute and relative imports in python scripts
Absolute and relative imports work when a file is executed as a module,
e.g. `python3 -m util.foo.bar`, but break when the same file is executed
as a script, e.g. `python3 util/foo/bar.py`. This commit adds the
`fix_imports` package that enables absolute and relative imports in
scripts.

Signed-off-by: Alphan Ulusoy <alphan@opentitan.org>